@@ -94,69 +137,72 @@
     "all": true
   },
   "engines": {
-    "node": ">= 4.1.2"
+    "node": ">=8.3.0"
   },
   "files": [
+    "build",
+    "bundles/.empty_directory",
+    "db/.empty_directory",
+    "lib",
+    "schemas",
+    "src",
+    "types",
     "AUTHORS",
-    "LICENSE",
-    "README.md",
-    "bower.json",
-    "index.js",
-    "lib/"
+    "package-lock.json"
   ],
   "pkg": {
     "assets": [
-      "bower_components/**/*",
       "src/**/*",
       "build/**/*",
       "bundles/**/*",
-      "schemas/**/*"
+      "schemas/**/*",
+      "node_modules/babel-plugin-bare-import-rewrite/**"
     ]
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.4.3",
     "aliasify": "^2.1.0",
-    "ava": "^0.25.0",
-    "axios": "^0.18.0",
-    "babel-plugin-istanbul": "^4.1.5",
-    "babel-preset-es2015-nostrict": "^6.6.2",
-    "babel-preset-minify": "^0.2.0",
-    "babelify": "^8.0.0",
-    "brfs": "^2.0.0",
-    "browserify": "^14.5.0",
-    "chai": "^4.1.2",
-    "del": "^3.0.0",
-    "eslint": "^4.10.0",
-    "eslint-config-xo": "^0.19.0",
-    "eslint-plugin-ava": "^4.2.2",
-    "exorcist": "^1.0.0",
-    "gh-pages": "^1.2.0",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^7.0.0",
-    "gulp-clean-css": "^3.9.0",
-    "gulp-htmlmin": "^4.0.0",
-    "gulp-if": "^2.0.2",
-    "gulp-imagemin": "^4.0.0",
-    "gulp-sourcemaps": "^2.6.4",
+    "ava": "^2.2.0",
+    "axios": "^0.18.1",
+    "babel-plugin-istanbul": "^5.1.1",
+    "babel-preset-minify": "^0.5.0",
+    "babelify": "^10.0.0",
+    "brfs": "^2.0.2",
+    "browserify": "^16.2.3",
+    "del": "^4.1.0",
+    "eslint": "^6.1.0",
+    "eslint-config-xo": "^0.26.0",
+    "eslint-plugin-ava": "^7.1.0",
+    "exorcist": "^1.0.1",
+    "fs-extra": "^7.0.1",
+    "gh-pages": "^2.0.1",
+    "gulp": "^4.0.0",
     "gulp-util": "^3.0.8",
-    "is-windows": "^1.0.1",
-    "jsdoc": "^3.5.5",
-    "merge-stream": "^1.0.1",
-    "minami": "github:nodecg/minami#a674ded7782ed7c60e8d5c1638aace339cbdbb4c",
-    "nyc": "^13.0.0",
-    "polymer-build": "^3.0.4",
-    "remark": "^8.0.0",
-    "remark-toc": "^5.0.0",
-    "replace-in-file": "^3.4.0",
-    "rimraf": "^2.6.2",
-    "simple-git": "^1.96.0",
-    "sinon": "^4.1.2",
-    "socket.io-client": "^2.1.1",
-    "standard-version": "^4.4.0",
+    "is-windows": "^1.0.2",
+    "jsdoc": "^3.6.2",
+    "minami": "github:nodecg/minami#fbe470b796dd26f2b3d04584e5fc90451bea3ab8",
+    "nyc": "^14.1.1",
+    "p-retry": "^4.1.0",
+    "puppeteer": "^1.19.0",
+    "remark": "^10.0.1",
+    "remark-toc": "^5.1.1",
+    "replace-in-file": "^3.4.4",
+    "rimraf": "^2.6.3",
+    "simple-git": "^1.110.0",
+    "sinon": "^7.3.1",
+    "socket.io-client": "^2.2.0",
+    "standard-version": "^6.0.1",
     "taffydb": "^2.7.3",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0",
+    "temp": "^0.9.0",
+    "vinyl-buffer": "^1.0.1",
+    "vinyl-source-stream": "^2.0.0",
     "vinyl-transform": "^1.0.0",
-    "wdio-webcomponents": "^1.1.0",
-    "webdriverio": "^4.13.1"
-  }
+    "wdio-webcomponents": "^1.1.0"
+  },
+  "browserslist": [
+    "last 2 chrome versions",
+    "last 2 firefox versions",
+    "last 2 safari versions",
+    "last 2 edge versions"
+  ]
 }
