@@ -1,20 +1,24 @@
 'use strict';
 
 // Packages
-const test = require('ava');
+import test from 'ava';
 
 // Ours
-require('../helpers/nodecg-and-webdriver')(test, {tabs: ['graphic']}); // Must be first.
-const e = require('../helpers/test-environment');
+import * as server from '../helpers/server';
+import * as browser from '../helpers/browser';
 
-test.beforeEach(async () => {
-	await e.browser.client.switchTab(e.browser.tabs.graphic);
+server.setup();
+const {initGraphic} = browser.setup();
+
+let graphic;
+test.before(async () => {
+	graphic = await initGraphic();
 });
 
 // The graphic and dashboard APIs use the same file
 // If dashboard API passes all its tests, we just need to make sure that the socket works
 test.serial('should receive messages', async t => {
-	await e.browser.client.execute(() => {
+	await graphic.evaluate(() => {
 		window.serverToGraphicReceived = false;
 		window.graphicApi.listenFor('serverToGraphic', () => {
 			window.serverToGraphicReceived = true;
@@ -22,39 +26,41 @@ test.serial('should receive messages', async t => {
 	});
 
 	const sendMessageInterval = setInterval(() => {
-		e.apis.extension.sendMessage('serverToGraphic');
+		t.context.apis.extension.sendMessage('serverToGraphic');
 	}, 500);
 
-	await e.browser.client.executeAsync(done => {
+	await graphic.evaluate(() => new Promise(resolve => {
 		const checkMessageReceived = setInterval(() => {
 			if (window.serverToGraphicReceived) {
 				clearInterval(checkMessageReceived);
-				done();
+				resolve();
 			}
 		}, 50);
-	});
+	}));
 
 	clearInterval(sendMessageInterval);
 	t.pass();
 });
 
-test.cb.serial('should send messages', t => {
-	e.apis.extension.listenFor('graphicToServer', t.end);
-	e.browser.client.execute(() => window.graphicApi.sendMessage('graphicToServer'));
+test.serial.cb('should send messages', t => {
+	t.context.apis.extension.listenFor('graphicToServer', t.end);
+	graphic.evaluate(() => {
+		window.graphicApi.sendMessage('graphicToServer');
+	});
 });
 
 test.serial('#bundleVersion', async t => {
-	const res = await e.browser.client.execute(() => {
+	const res = await graphic.evaluate(() => {
 		return window.graphicApi.bundleVersion;
 	});
-	t.is(res.value, '0.0.1');
+	t.is(res, '0.0.1');
 });
 
 test.serial('#bundleGit', async t => {
-	const res = await e.browser.client.execute(() => {
+	const res = await graphic.evaluate(() => {
 		return window.graphicApi.bundleGit;
 	});
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		branch: 'master',
 		date: '2018-07-13T17:09:29.000Z',
 		hash: '6262681c7f35eccd7293d57a50bdd25e4cd90684',
