@@ -1,63 +1,66 @@
 'use strict';
 
 // Packages
-const test = require('ava');
+import test from 'ava';
+import * as axios from 'axios';
 const express = require('express');
-const axios = require('axios');
 
 // Ours
-require('../helpers/nodecg-and-webdriver')(test, {tabs: ['dashboard']}); // Must be first.
-const C = require('../helpers/test-constants');
-const e = require('../helpers/test-environment');
+import * as server from '../helpers/server';
+import * as browser from '../helpers/browser';
+
+server.setup();
+const {initDashboard} = browser.setup();
+
+import * as C from '../helpers/test-constants';
+
+let dashboard;
+test.before(async () => {
+	dashboard = await initDashboard();
+});
 
 test.serial('should receive messages and fire acknowledgements', async t => {
-	e.apis.extension.listenFor('clientToServer', (data, cb) => cb());
-	await e.browser.client.executeAsync(done => {
-		return window.dashboardApi.sendMessage('clientToServer', null, done);
-	});
+	t.context.apis.extension.listenFor('clientToServer', (data, cb) => cb());
+	await dashboard.evaluate(() => new Promise(resolve => {
+		window.dashboardApi.sendMessage('clientToServer', null, resolve);
+	}));
 	t.pass();
 });
 
 test.serial('should serialize errors sent to acknowledgements', async t => {
-	e.apis.extension.listenFor('ackErrors', (data, cb) => cb(new Error('boom')));
-	const res = await e.browser.client.executeAsync(done => {
+	t.context.apis.extension.listenFor('ackErrors', (data, cb) => cb(new Error('boom')));
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		window.dashboardApi.sendMessage('ackErrors', null, err => {
-			done(err.message);
+			resolve(err.message);
 		});
-	});
-	t.is(res.value, 'boom');
+	}));
+	t.is(res, 'boom');
 });
 
 test.serial('should resolve acknowledgement promises', async t => {
-	e.apis.extension.listenFor('ackPromiseResolve', (data, cb) => cb());
-	const res = await e.browser.client.executeAsync(done => {
-		window.dashboardApi.sendMessage('ackPromiseResolve').then(() => {
-			done();
-		}).catch(done);
-	});
-	t.is(res.value, null);
+	t.context.apis.extension.listenFor('ackPromiseResolve', (data, cb) => cb());
+	const res = await dashboard.evaluate(() =>
+		window.dashboardApi.sendMessage('ackPromiseResolve').catch());
+	t.is(res, undefined);
 });
 
 test.serial('should reject acknowledgement promises if there was an error', async t => {
-	e.apis.extension.listenFor('ackPromiseReject', (data, cb) => cb(new Error('boom')));
-	const res = await e.browser.client.executeAsync(done => {
-		window.dashboardApi.sendMessage('ackPromiseReject').then(() => {
-			done(new Error('Promise resolved when it should have rejected.'));
-		}).catch(err => {
-			done(err.message);
-		});
-	});
-	t.is(res.value, 'boom');
+	t.context.apis.extension.listenFor('ackPromiseReject', (data, cb) => cb(new Error('boom')));
+	const res = await dashboard.evaluate(() =>
+		window.dashboardApi.sendMessage('ackPromiseReject')
+			.then(() => new Error('Promise resolved when it should have rejected.'))
+			.catch(err => err.message));
+	t.is(res, 'boom');
 });
 
 test.serial('should not return a promise if the user provided a callback ', async t => {
-	e.apis.extension.listenFor('ackPromiseCallback', (data, cb) => cb());
-	const res = await e.browser.client.executeAsync(done => {
+	t.context.apis.extension.listenFor('ackPromiseCallback', (data, cb) => cb());
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const returnVal = window.dashboardApi.sendMessage('ackPromiseCallback', () => {
-			done(returnVal === undefined);
+			resolve(returnVal === undefined);
 		});
-	});
-	t.true(res.value);
+	}));
+	t.true(res);
 });
 
 test('should mount custom routes via nodecg.mount', async t => {
@@ -65,9 +68,9 @@ test('should mount custom routes via nodecg.mount', async t => {
 	app.get('/test-bundle/test-route', (req, res) => {
 		res.send('custom route confirmed');
 	});
-	e.apis.extension.mount(app);
+	t.context.apis.extension.mount(app);
 
-	const response = await axios.get(`${C.ROOT_URL}test-bundle/test-route`);
+	const response = await axios.get(`${C.rootUrl()}test-bundle/test-route`);
 	t.is(response.status, 200);
 	t.is(response.data, 'custom route confirmed');
 });
@@ -76,18 +79,18 @@ test.serial.cb('should support multiple listenFor handlers', t => {
 	t.plan(2);
 	let callbacksInvoked = 0;
 
-	e.apis.extension.listenFor('multipleListenFor', () => {
+	t.context.apis.extension.listenFor('multipleListenFor', () => {
 		t.pass();
 		checkDone();
 	});
 
-	e.apis.extension.listenFor('multipleListenFor', () => {
+	t.context.apis.extension.listenFor('multipleListenFor', () => {
 		t.pass();
 		checkDone();
 	});
 
-	e.browser.client.execute(() => {
-		return window.dashboardApi.sendMessage('multipleListenFor');
+	dashboard.evaluate(() => {
+		window.dashboardApi.sendMessage('multipleListenFor');
 	});
 
 	function checkDone() {
@@ -102,21 +105,21 @@ test.serial.cb('should prevent acknowledgements from being called more than once
 	t.plan(4);
 	let callbacksInvoked = 0;
 
-	e.apis.extension.listenFor('singleAckEnforcement', (data, cb) => {
+	t.context.apis.extension.listenFor('singleAckEnforcement', (data, cb) => {
 		t.false(cb.handled);
 		t.notThrows(cb);
 		checkDone();
 	});
 
-	e.apis.extension.listenFor('singleAckEnforcement', (data, cb) => {
+	t.context.apis.extension.listenFor('singleAckEnforcement', (data, cb) => {
 		t.true(cb.handled);
 		t.throws(cb);
 		checkDone();
 	});
 
-	e.browser.client.executeAsync(done => {
-		return window.dashboardApi.sendMessage('singleAckEnforcement', null, done);
-	});
+	dashboard.evaluate(() => new Promise(resolve => {
+		return window.dashboardApi.sendMessage('singleAckEnforcement', null, resolve);
+	}));
 
 	function checkDone() {
 		callbacksInvoked++;
@@ -126,67 +129,80 @@ test.serial.cb('should prevent acknowledgements from being called more than once
 	}
 });
 
-test.serial('server - should support intra-context messaging', async t => {
+test.serial.cb('server - should support intra-context messaging', t => {
 	t.plan(2);
 
-	// This is what we're actually testing.
-	e.apis.extension.listenFor('serverToServer', data => {
-		t.deepEqual(data, {foo: 'bar'});
-	});
-
-	// But, we also make sure that the client (browser) is still getting these messages as normal.
-	await e.browser.client.execute(() => {
-		window.dashboardApi.listenFor('serverToServer', data => {
-			window._serverToServerData = data;
-		});
-	});
-
-	// Send the message only after both listeners have been set up.
-	e.apis.extension.sendMessage('serverToServer', {foo: 'bar'});
-
-	// Wait until the browser has received the message.
-	await e.browser.client.waitUntil(async () => {
-		const response = await e.browser.client.execute(() => {
-			return window._serverToServerData;
+	let assertCount = 0;
+	(async () => {
+		// This is what we're actually testing.
+		t.context.apis.extension.listenFor('serverToServer', data => {
+			t.deepEqual(data, {foo: 'bar'});
+			if (++assertCount === 2) {
+				t.end();
+			}
 		});
 
-		return response && response.value &&
-			typeof response.value === 'object' &&
-			Object.keys(response.value).length > 0;
-	});
-
-	// Verify that the browser got the right data along with the message.
-	const response = await e.browser.client.execute(() => {
-		return window._serverToServerData;
-	});
-	t.deepEqual(response.value, {foo: 'bar'});
-});
-
-test.serial('client - should support intra-context messaging', async t => {
-	t.plan(2);
-
-	// We also want to make sure that the server (extension) is still getting these messages as normal.
-	e.apis.extension.listenFor('clientToClient', data => {
-		t.deepEqual(data, {baz: 'qux'});
-	});
-
-	const response = await e.browser.client.executeAsync(done => {
-		window.dashboardApi.listenFor('clientToClient', data => {
-			done(data);
+		// But, we also make sure that the client (browser) is still getting these messages as normal.
+		await dashboard.evaluate(() => {
+			window.dashboardApi.listenFor('serverToServer', data => {
+				window._serverToServerData = data;
+			});
 		});
 
 		// Send the message only after both listeners have been set up.
-		window.dashboardApi.sendMessage('clientToClient', {baz: 'qux'});
-	});
-	t.deepEqual(response.value, {baz: 'qux'});
+		t.context.apis.extension.sendMessage('serverToServer', {foo: 'bar'});
+
+		// Wait until the browser has received the message.
+		await dashboard.waitForFunction(async () => {
+			const data = window._serverToServerData;
+			return typeof data === 'object' && Object.keys(data).length > 0;
+		});
+
+		// Verify that the browser got the right data along with the message.
+		const response = await dashboard.evaluate(() => {
+			return window._serverToServerData;
+		});
+		t.deepEqual(response, {foo: 'bar'});
+		if (++assertCount === 2) {
+			t.end();
+		}
+	})();
+});
+
+test.serial.cb('client - should support intra-context messaging', t => {
+	t.plan(2);
+
+	let asserCount = 0;
+	(async () => {
+		// We also want to make sure that the server (extension) is still getting these messages as normal.
+		t.context.apis.extension.listenFor('clientToClient', data => {
+			t.deepEqual(data, {baz: 'qux'});
+			if (++asserCount === 2) {
+				t.end();
+			}
+		});
+
+		const response = await dashboard.evaluate(() => new Promise(resolve => {
+			window.dashboardApi.listenFor('clientToClient', data => {
+				resolve(data);
+			});
+
+			// Send the message only after both listeners have been set up.
+			window.dashboardApi.sendMessage('clientToClient', {baz: 'qux'});
+		}));
+		t.deepEqual(response, {baz: 'qux'});
+		if (++asserCount === 2) {
+			t.end();
+		}
+	})();
 });
 
 test('server - #bundleVersion', t => {
-	t.is(e.apis.extension.bundleVersion, '0.0.1');
+	t.is(t.context.apis.extension.bundleVersion, '0.0.1');
 });
 
 test('server - #bundleGit', t => {
-	t.deepEqual(e.apis.extension.bundleGit, {
+	t.deepEqual(t.context.apis.extension.bundleGit, {
 		branch: 'master',
 		date: new Date('2018-07-13T17:09:29.000Z'),
 		hash: '6262681c7f35eccd7293d57a50bdd25e4cd90684',
@@ -196,6 +212,6 @@ test('server - #bundleGit', t => {
 });
 
 test('bundles replicant', t => {
-	const bundlesRep = e.apis.extension.Replicant('bundles', 'nodecg');
+	const bundlesRep = t.context.apis.extension.Replicant('bundles', 'nodecg');
 	t.is(bundlesRep.value.length, 5);
 });
