@@ -1,14 +1,14 @@
 'use strict';
 
 // Native
-const fs = require('fs');
-const path = require('path');
+import * as fs from 'fs';
+import * as path from 'path';
 
 // Packages
-const fse = require('fs-extra');
+import * as fse from 'fs-extra';
+import * as temp from 'temp';
+import * as test from 'ava';
 const isWindows = require('is-windows');
-const temp = require('temp');
-const test = require('ava');
 
 // Ours
 const Logger = require('../lib/logger/server')({console: {enabled: true}});
@@ -17,7 +17,7 @@ const tempFolder = temp.mkdirSync();
 temp.track(); // Automatically track and cleanup files at exit.
 
 let bundleManager;
-test.cb.before(t => {
+test.before.cb(t => {
 	fse.copySync('test/fixtures/bundle-manager', tempFolder);
 
 	// The symlink test can't run on Windows unless run with admin privs.
@@ -39,7 +39,7 @@ test.cb.before(t => {
 
 	bundleManager = require('../lib/bundle-manager');
 	bundleManager.init(
-		path.join(tempFolder, 'bundles'),
+		[path.join(tempFolder, 'bundles'), path.join(tempFolder, 'custom-bundles')],
 		path.join(tempFolder, 'cfg'),
 		'0.7.0',
 		nodecgConfig,
@@ -77,7 +77,12 @@ test.serial('loader - should not crash or load an invalid bundle', t => {
 	t.is(bundle, undefined);
 });
 
-test.cb.serial('watcher - hould emit a change event when the manifest file changes', t => {
+test.serial('loader - should detect and load bundle located in custom bundle paths', t => {
+	const bundle = bundleManager.find('another-test-bundle');
+	t.is(bundle.name, 'another-test-bundle');
+});
+
+test.serial.cb('watcher - hould emit a change event when the manifest file changes', t => {
 	const manifest = JSON.parse(fs.readFileSync(`${tempFolder}/bundles/change-manifest/package.json`));
 	bundleManager.once('bundleChanged', bundle => {
 		t.is(bundle.name, 'change-manifest');
@@ -88,7 +93,7 @@ test.cb.serial('watcher - hould emit a change event when the manifest file chang
 	fs.writeFileSync(`${tempFolder}/bundles/change-manifest/package.json`, JSON.stringify(manifest));
 });
 
-test.cb.serial('watcher - should remove the bundle when the manifest file is renamed', t => {
+test.serial.cb('watcher - should remove the bundle when the manifest file is renamed', t => {
 	bundleManager.once('bundleRemoved', () => {
 		const result = bundleManager.find('rename-manifest');
 		t.is(result, undefined);
@@ -101,7 +106,7 @@ test.cb.serial('watcher - should remove the bundle when the manifest file is ren
 	);
 });
 
-test.cb.serial('watcher - should emit a removed event when the manifest file is removed', t => {
+test.serial.cb('watcher - should emit a removed event when the manifest file is removed', t => {
 	bundleManager.once('bundleRemoved', () => {
 		const result = bundleManager.find('remove-manifest');
 		t.is(result, undefined);
@@ -111,7 +116,7 @@ test.cb.serial('watcher - should emit a removed event when the manifest file is 
 	fs.unlinkSync(`${tempFolder}/bundles/remove-manifest/package.json`);
 });
 
-test.cb.serial('watcher - should emit a change event when a panel HTML file changes', t => {
+test.serial.cb('watcher - should emit a change event when a panel HTML file changes', t => {
 	bundleManager.once('bundleChanged', bundle => {
 		t.is(bundle.name, 'change-panel');
 		t.end();
@@ -126,7 +131,7 @@ test.cb.serial('watcher - should emit a change event when a panel HTML file chan
 if (!isWindows()) {
 	// This can't be tested on Windows unless run with admin privs.
 	// For some reason, creating symlinks on Windows requires admin.
-	test.cb.serial('watcher - should detect panel HTML file changes when the bundle is symlinked', t => {
+	test.serial.cb('watcher - should detect panel HTML file changes when the bundle is symlinked', t => {
 		bundleManager.once('bundleChanged', bundle => {
 			t.is(bundle.name, 'change-panel-symlink');
 			t.end();
@@ -139,7 +144,7 @@ if (!isWindows()) {
 	});
 }
 
-test.cb.serial('watcher - should reload the bundle\'s config when the bundle is reloaded due to a change', t => {
+test.serial.cb('watcher - should reload the bundle\'s config when the bundle is reloaded due to a change', t => {
 	const manifest = JSON.parse(fs.readFileSync(`${tempFolder}/bundles/change-config/package.json`));
 	const config = JSON.parse(fs.readFileSync(`${tempFolder}/cfg/change-config.json`));
 
@@ -158,7 +163,7 @@ test.cb.serial('watcher - should reload the bundle\'s config when the bundle is 
 	fs.writeFileSync(`${tempFolder}/cfg/change-config.json`, JSON.stringify(config));
 });
 
-test.cb.serial('watcher - should emit an `invalidBundle` error when a panel HTML file is removed', t => {
+test.serial.cb('watcher - should emit an `invalidBundle` error when a panel HTML file is removed', t => {
 	bundleManager.once('invalidBundle', (bundle, error) => {
 		t.is(bundle.name, 'remove-panel');
 		t.is(error.message, 'Panel file "panel.html" in bundle "remove-panel" does not exist.');
@@ -168,7 +173,7 @@ test.cb.serial('watcher - should emit an `invalidBundle` error when a panel HTML
 	fs.unlinkSync(`${tempFolder}/bundles/remove-panel/dashboard/panel.html`);
 });
 
-test.cb.serial('watcher - should emit an `invalidBundle` error when the manifest becomes invalid', t => {
+test.serial.cb('watcher - should emit an `invalidBundle` error when the manifest becomes invalid', t => {
 	bundleManager.once('invalidBundle', (bundle, error) => {
 		t.is(bundle.name, 'invalid-manifest');
 		t.is(error.message, `${path.join(tempFolder, 'bundles/invalid-manifest/package.json')} is not valid JSON, please check it against a validator such as jsonlint.com`);
