@@ -1,12 +1,14 @@
 'use strict';
 
 // Packages
-const test = require('ava');
-const axios = require('axios');
+import test from 'ava';
+import * as axios from 'axios';
 
 // Ours
-require('./helpers/nodecg-and-webdriver')(test); // Must be first.
-const C = require('./helpers/test-constants');
+import * as server from './helpers/server';
+server.setup();
+
+import * as C from './helpers/test-constants';
 const bundleManager = require('../lib/bundle-manager');
 
 test('should load bundles which have satisfied bundle dependencies', t => {
@@ -22,20 +24,20 @@ test('should not load bundles which have unsatisfied bundle dependencies', t => 
 });
 
 test('should serve bundle-specific bower_components', async t => {
-	const response = await axios.get(`${C.BUNDLE_BOWER_COMPONENTS_URL}confirmation.js`);
+	const response = await axios.get(`${C.bundleBowerComponentsUrl()}confirmation.js`);
 	t.is(response.status, 200);
 	t.is(response.data, 'const confirmed = \'bower_components_confirmed\';\n');
 });
 
 test('should serve bundle-specific node_modules', async t => {
-	const response = await axios.get(`${C.BUNDLE_NODE_MODULES_URL}confirmation.js`);
+	const response = await axios.get(`${C.bundleNodeModulesUrl()}confirmation.js`);
 	t.is(response.status, 200);
 	t.is(response.data, 'const confirmed = \'node_modules_confirmed\';\n');
 });
 
 test('should 404 on non-existent bower_component', async t => {
 	try {
-		await axios.get(`${C.BUNDLE_BOWER_COMPONENTS_URL}confirmation_404.js`);
+		await axios.get(`${C.bundleBowerComponentsUrl()}confirmation_404.js`);
 	} catch (error) {
 		t.is(error.response.status, 404);
 	}
@@ -43,7 +45,7 @@ test('should 404 on non-existent bower_component', async t => {
 
 test('should 404 on non-existent node_module', async t => {
 	try {
-		await axios.get(`${C.BUNDLE_NODE_MODULES_URL}confirmation_404.js`);
+		await axios.get(`${C.bundleNodeModulesUrl()}confirmation_404.js`);
 	} catch (error) {
 		t.is(error.response.status, 404);
 	}
@@ -51,20 +53,20 @@ test('should 404 on non-existent node_module', async t => {
 
 test('should 404 on non-existent bundle node_modules/bower_components', async t => {
 	try {
-		await axios.get(`${C.ROOT_URL}bundles/false-bundle/node_modules/confirmation_404.js`);
+		await axios.get(`${C.rootUrl()}bundles/false-bundle/node_modules/confirmation_404.js`);
 	} catch (error) {
 		t.is(error.response.status, 404);
 	}
 });
 
 test('should redirect /login/ to /dashboard/ when login security is disabled', async t => {
-	const response = await axios.get(C.LOGIN_URL);
+	const response = await axios.get(C.loginUrl());
 	t.is(response.status, 200);
-	t.is(response.request.res.responseUrl, C.DASHBOARD_URL);
+	t.is(response.request.res.responseUrl, C.dashboardUrl());
 });
 
 test.serial('shared sources - 200', async t => {
-	const response = await axios.get(`${C.ROOT_URL}bundles/test-bundle/shared/util.js`);
+	const response = await axios.get(`${C.rootUrl()}bundles/test-bundle/shared/util.js`);
 	t.is(response.status, 200);
 	t.is(
 		response.data,
@@ -77,7 +79,7 @@ test.serial('shared sources - 200', async t => {
 
 test('shared sources - 404', async t => {
 	try {
-		await axios.get(`${C.ROOT_URL}bundles/test-bundle/shared/404.js`);
+		await axios.get(`${C.rootUrl()}bundles/test-bundle/shared/404.js`);
 	} catch (error) {
 		t.is(error.response.status, 404);
 	}
@@ -85,7 +87,7 @@ test('shared sources - 404', async t => {
 
 test('shared sources - no bundle 404', async t => {
 	try {
-		await axios.get(`${C.ROOT_URL}bundles/false-bundle/shared/404.js`);
+		await axios.get(`${C.rootUrl()}bundles/false-bundle/shared/404.js`);
 	} catch (error) {
 		t.is(error.response.status, 404);
 	}
