@@ -1,19 +1,20 @@
 'use strict';
 
 // Native
-const fs = require('fs');
-const path = require('path');
+import * as fs from 'fs';
+import * as path from 'path';
 
 // Packages
-const test = require('ava');
+import test from 'ava';
 const fetch = require('make-fetch-happen');
 
 // Ours
-require('./helpers/nodecg-and-webdriver')(test); // Must be first.
-const C = require('./helpers/test-constants');
+import * as server from './helpers/server';
+server.setup();
+import * as C from './helpers/test-constants';
 
 test('serves files from custom mountpoints', async t => {
-	const response = await fetch(`${C.ROOT_URL}bundles/test-bundle/custom-mount/hello-world.html`);
+	const response = await fetch(`${C.rootUrl()}bundles/test-bundle/custom-mount/hello-world.html`);
 	const bodyText = await response.text();
 
 	const filePath = path.resolve(
@@ -24,6 +25,6 @@ test('serves files from custom mountpoints', async t => {
 });
 
 test('returns a 404 when the file is not found', async t => {
-	const response = await fetch(`${C.ROOT_URL}bundles/test-bundle/custom-mount/not-found.html`);
+	const response = await fetch(`${C.rootUrl()}bundles/test-bundle/custom-mount/not-found.html`);
 	t.is(response.status, 404);
 });
