@@ -1,39 +1,55 @@
 'use strict';
 
 // Native
-const fs = require('fs');
-const path = require('path');
+import * as fs from 'fs';
+import * as path from 'path';
 
 // Packages
-const axios = require('axios');
+import test from 'ava';
+import * as axios from 'axios';
 const simpleGit = require('simple-git/promise');
 const replace = require('replace-in-file');
-const test = require('ava');
 
 // Ours
-require('./helpers/nodecg-and-webdriver')(test, {tabs: ['dashboard', 'single-instance']}); // Must be first.
-const C = require('./helpers/test-constants');
-const e = require('./helpers/test-environment');
-const util = require('./helpers/utilities');
+import * as server from './helpers/server';
+import * as browser from './helpers/browser';
+server.setup();
+const {initSingleInstance, initDashboard, initGraphic} = browser.setup();
+import * as C from './helpers/test-constants';
+import * as util from './helpers/utilities';
 
+let singleInstance;
+let dashboard;
 test.before(async () => {
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await e.browser.client.click('ncg-dashboard paper-tab[data-route="graphics"]');
-	await e.browser.client.click('ncg-dashboard ncg-graphics ncg-graphics-bundle ncg-graphic #collapseButton');
-});
+	singleInstance = await initSingleInstance();
+	dashboard = await initDashboard();
 
-test.beforeEach(async () => {
-	await e.browser.client.switchTab(e.browser.tabs.singleInstance);
+	const graphicButton = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'paper-tab[data-route="graphics"]'
+	);
+	await graphicButton.click();
+
+	const collapseButton = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+		'ncg-graphic',
+		'#collapseButton'
+	);
+	await collapseButton.click();
 });
 
 test('singleInstance - scripts get injected into /instance/*.html routes', async t => {
-	const response = await axios.get(`${C.ROOT_URL}instance/killed.html`);
+	const response = await axios.get(`${C.rootUrl()}instance/killed.html`);
 	t.is(response.status, 200);
 	t.true(response.data.includes('<script src="/nodecg-api.min.js">'));
 	t.true(response.data.includes('<script src="/socket.io/socket.io.js"></script>'));
 });
 
-test.serial.cb('singleInstance - shouldn\'t enter an infinite redirect loop when including a polymer element that loads an external stylesheet', t => {
+test.cb('singleInstance - shouldn\'t enter an infinite redirect loop when including a polymer element that loads an external stylesheet', t => {
 	const registration = require('../lib/graphics/registration');
 
 	function cb(url) {
@@ -53,96 +69,112 @@ test.serial.cb('singleInstance - shouldn\'t enter an infinite redirect loop when
 });
 
 test.serial('singleInstance - should redirect to busy.html when the instance is already taken', async t => {
-	await e.browser.client.newWindow(C.SINGLE_INSTANCE_URL);
-	await util.sleep(1000);
-	t.is(
-		await e.browser.client.getUrl(),
-		`${C.ROOT_URL}instance/busy.html?pathname=/bundles/test-bundle/graphics/single_instance.html`
-	);
+	t.plan(0);
+	const page = await initSingleInstance();
+	await page.waitForFunction(rootUrl =>
+		location.href === `${rootUrl}instance/busy.html?pathname=/bundles/test-bundle/graphics/single_instance.html`, {}, C.rootUrl());
 });
 
 test.serial('singleInstance - should redirect to killed.html when the instance is killed', async t => {
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await e.browser.client.execute(() => {
-		document.querySelector('ncg-dashboard').shadowRoot
-			.querySelector('ncg-graphics').shadowRoot
-			.querySelector('ncg-graphics-bundle').shadowRoot
-			.querySelectorAll('ncg-graphic')[1].shadowRoot
-			.querySelector('ncg-graphic-instance').$.killButton.click();
-	});
+	t.plan(0);
 
-	await e.browser.client.switchTab(e.browser.tabs.singleInstance);
-	t.is(
-		await e.browser.client.getUrl(),
-		`${C.ROOT_URL}instance/killed.html?pathname=/bundles/test-bundle/graphics/single_instance.html`
+	const graphicBoard = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+		'ncg-graphic:nth-of-type(2)',
 	);
+
+	await dashboard.bringToFront();
+	const expandButton = await dashboard.evaluateHandle(el => el.shadowRoot.querySelector('paper-button#collapseButton'), graphicBoard);
+	await expandButton.click();
+
+	const button = await dashboard.evaluateHandle(el =>
+		el.shadowRoot.querySelector('ncg-graphic-instance').$.killButton, graphicBoard);
+	await button.click();
+
+	await singleInstance.waitForFunction(rootUrl =>
+		location.href === `${rootUrl}instance/killed.html?pathname=/bundles/test-bundle/graphics/single_instance.html`, {}, C.rootUrl());
 });
 
 test.serial('singleInstance - should allow the graphic to be taken after being killed', async t => {
-	await e.browser.client.newWindow(C.SINGLE_INSTANCE_URL);
-	await util.sleep(500);
-	t.is(await e.browser.client.getUrl(), C.SINGLE_INSTANCE_URL);
+	const page = await initSingleInstance();
+	t.is(page.url(), C.singleInstanceUrl());
 });
 
 test.serial('refresh all instances in a bundle', async t => {
-	await e.browser.client.newWindow(C.GRAPHIC_URL);
-	const graphicTabId = await e.browser.client.getCurrentTabId();
-	await util.waitForRegistration();
+	const graphic = await initGraphic();
+	await util.waitForRegistration(graphic);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await util.sleep(50);
-	await e.browser.client.execute(() => {
-		const graphicsBundleEl = document.querySelector('ncg-dashboard').shadowRoot
-			.querySelector('ncg-graphics').shadowRoot
-			.querySelector('ncg-graphics-bundle');
-		graphicsBundleEl.$.reloadButton.click();
-		graphicsBundleEl.shadowRoot.querySelector('paper-button[dialog-confirm]').click();
-	});
+	const graphicBundle = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+	);
 
-	await e.browser.client.switchTab(graphicTabId);
-	await util.sleep(500);
-	const refreshMarker = await util.waitForRegistration();
-	t.is(refreshMarker, null);
+	await dashboard.bringToFront();
+	const reload = await dashboard.evaluateHandle(el => el.$.reloadButton, graphicBundle);
+	await reload.click();
+	const confirm = await dashboard.evaluateHandle(el => el.shadowRoot.querySelector('paper-button[dialog-confirm]'), graphicBundle);
+	await confirm.click();
+
+	await graphic.waitFor(500);
+	const refreshMarker = await util.waitForRegistration(graphic);
+	t.is(refreshMarker, undefined);
 });
 
 test.serial('refresh all instances of a graphic', async t => {
-	await e.browser.client.newWindow(C.GRAPHIC_URL);
-	const graphicTabId = await e.browser.client.getCurrentTabId();
-	await util.waitForRegistration();
+	const graphic = await initGraphic();
+	await util.waitForRegistration(graphic);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await util.sleep(50);
-	await e.browser.client.click('ncg-dashboard ncg-graphics ncg-graphics-bundle' +
-		' ncg-graphic #reloadButton');
+	await dashboard.bringToFront();
+	const reload = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+		'ncg-graphic',
+		'#reloadButton',
+	);
+	await reload.click();
 
-	await e.browser.client.switchTab(graphicTabId);
-	await util.sleep(500);
-	const refreshMarker = await util.waitForRegistration();
-	await e.browser.client.close(); // We don't need the coverage data from this tab.
-	t.is(refreshMarker, null);
+	await graphic.waitFor(500);
+	const refreshMarker = await util.waitForRegistration(graphic);
+	t.is(refreshMarker, undefined);
 });
 
 test.serial('refresh individual instance', async t => {
-	await e.browser.client.newWindow(C.GRAPHIC_URL);
-	const graphicTabId = await e.browser.client.getCurrentTabId();
-	await util.waitForRegistration();
+	const graphic = await initGraphic();
+	await util.waitForRegistration(graphic);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await util.sleep(50);
-	await e.browser.client.execute(() => {
-		document.querySelector('ncg-dashboard').shadowRoot
-			.querySelector('ncg-graphics').shadowRoot
-			.querySelector('ncg-graphics-bundle').shadowRoot
-			.querySelectorAll('ncg-graphic')[0].shadowRoot
-			.querySelector('ncg-graphic-instance:last-of-type').$.reloadButton.click();
-	});
+	await dashboard.bringToFront();
+	const reload = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+		'ncg-graphic',
+		'ncg-graphic-instance:last-of-type',
+		'#reloadButton',
+	);
+	await reload.click();
 
-	await e.browser.client.switchTab(graphicTabId);
-	await util.sleep(1000);
-	const refreshMarker = await util.waitForRegistration();
-	await e.browser.client.close(); // We don't need the coverage data from this tab.
-	t.is(refreshMarker, null);
+	await graphic.waitFor(500);
+	const refreshMarker = await util.waitForRegistration(graphic);
+	t.is(refreshMarker, undefined);
 });
+
+const statusEl = page => util.shadowSelector(
+	page,
+	'ncg-dashboard',
+	'ncg-graphics',
+	'ncg-graphics-bundle',
+	'ncg-graphic',
+	'ncg-graphic-instance:last-of-type',
+	'#status'
+);
 
 test.serial('version out of date', async t => {
 	replace.sync({
@@ -150,11 +182,9 @@ test.serial('version out of date', async t => {
 		from: '"version": "0.0.1"',
 		to: '"version": "0.0.2"'
 	});
-	await util.sleep(1500);
+	await dashboard.waitFor(1500);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	let text = await e.browser.client.getText('ncg-dashboard ncg-graphics ncg-graphics-bundle ' +
-		'ncg-graphic ncg-graphic-instance:last-of-type #status');
+	let text = await dashboard.evaluate(el => el.textContent, await statusEl(dashboard));
 	t.is(text, 'Potentially Out of Date');
 
 	replace.sync({
@@ -162,11 +192,9 @@ test.serial('version out of date', async t => {
 		from: '"version": "0.0.2"',
 		to: '"version": "0.0.1"'
 	});
-	await util.sleep(1500);
+	await dashboard.waitFor(1500);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	text = await e.browser.client.getText('ncg-dashboard ncg-graphics ncg-graphics-bundle ' +
-		'ncg-graphic ncg-graphic-instance:last-of-type #status');
+	text = await dashboard.evaluate(el => el.textContent, await statusEl(dashboard));
 	t.is(text, 'Latest');
 });
 
@@ -178,23 +206,37 @@ test.serial('git out of date', async t => {
 	const git = simpleGit(path.resolve(process.env.NODECG_ROOT, 'bundles/test-bundle'));
 	await git.add('./new_file.txt');
 	await git.commit('new commit');
-	await util.sleep(1500);
+	await dashboard.waitFor(1500);
 
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	const text = await e.browser.client.getText('ncg-dashboard ncg-graphics ncg-graphics-bundle ' +
-		'ncg-graphic ncg-graphic-instance:last-of-type #status');
+	const text = await dashboard.evaluate(el => el.textContent, await statusEl(dashboard));
 	t.is(text, 'Potentially Out of Date');
 });
 
 test.serial('shows a diff when hovering over "potentially out of date" status', async t => {
-	await e.browser.client.switchTab(e.browser.tabs.dashboard);
-	await e.browser.client.moveToObject('ncg-dashboard ncg-graphics ncg-graphics-bundle ' +
-		'ncg-graphic ncg-graphic-instance[status="out-of-date"] #status');
+	await dashboard.bringToFront();
+	const graphicInstance = await util.shadowSelector(
+		dashboard,
+		'ncg-dashboard',
+		'ncg-graphics',
+		'ncg-graphics-bundle',
+		'ncg-graphic',
+		'ncg-graphic-instance[status="out-of-date"]'
+	);
 
-	const diffSelector = 'ncg-dashboard ncg-graphics ncg-graphics-bundle ' +
-		'ncg-graphic ncg-graphic-instance[status="out-of-date"] #diff';
-	await e.browser.client.isVisible(diffSelector);
-	const diffText = await e.browser.client.getText(diffSelector);
+	await graphicInstance.hover();
+	await dashboard.waitForFunction(
+		el => getComputedStyle(el).display !== 'none',
+		{},
+		graphicInstance
+	);
+	const diffText = await dashboard.evaluate(el => el.$.diff.$.body.textContent, graphicInstance);
+	t.true(diffText.includes('Current:'));
+	t.true(diffText.includes('Latest:'));
+	t.regex(diffText, /0\.0\.1 - \w{7} \[Initial commit\]/);
+	t.regex(diffText, /0\.0\.1 - \w{7} \[new commit\]/);
 
-	t.true(/Current: 0\.0\.1 - 6262681 \[Initial commit]\nLatest: {2}0\.0\.1 - .{7} \[new commit]/.test(diffText));
+	const closeButton = await dashboard.evaluateHandle(el =>
+		el.$.diff.shadowRoot.querySelector('paper-icon-button'), graphicInstance);
+	await closeButton.click();
+	await dashboard.waitForFunction(el => getComputedStyle(el.$.diff).opacity === '0', {timeout: 100000}, graphicInstance);
 });
