@@ -1,22 +1,27 @@
 'use strict';
 
 // Packages
-const test = require('ava');
+import test from 'ava';
 
 // Ours
-require('../helpers/nodecg-and-webdriver')(test, {tabs: ['dashboard']}); // Must be first.
-const e = require('../helpers/test-environment');
-test.beforeEach(() => {
-	return e.browser.client.switchTab(e.browser.tabs.dashboard);
+import * as server from '../helpers/server';
+import * as browser from '../helpers/browser';
+
+server.setup();
+const {initDashboard} = browser.setup();
+
+let dashboard;
+test.before(async () => {
+	dashboard = await initDashboard();
 });
 
 test.serial('should create a default value based on the schema, if none is provided', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaDefaults');
-		rep.on('declared', () => done(rep.value));
-	});
+		rep.on('declared', () => resolve(JSON.parse(JSON.stringify(rep.value))));
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		object: {
 			numA: 0
@@ -25,7 +30,7 @@ test.serial('should create a default value based on the schema, if none is provi
 });
 
 test.serial('should accept the defaultValue when it passes validation', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaDefaultValuePass', {
 			defaultValue: {
 				string: 'foo',
@@ -34,10 +39,10 @@ test.serial('should accept the defaultValue when it passes validation', async t 
 				}
 			}
 		});
-		rep.on('declared', () => done(rep.value));
-	});
+		rep.on('declared', () => resolve(JSON.parse(JSON.stringify(rep.value))));
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: 'foo',
 		object: {
 			numA: 1
@@ -46,7 +51,7 @@ test.serial('should accept the defaultValue when it passes validation', async t 
 });
 
 test.serial('should throw when defaultValue fails validation', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaDefaultValueFail', {
 			defaultValue: {
 				string: 0
@@ -54,23 +59,23 @@ test.serial('should throw when defaultValue fails validation', async t => {
 		});
 
 		rep.once('declarationRejected', reason => {
-			done(reason);
+			resolve(reason);
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaDefaultValueFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaDefaultValueFail"'));
 });
 
 test.serial('should accept the persisted value when it passes validation', async t => {
 	// Persisted value is copied from fixtures
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaPersistencePass');
 		rep.on('declared', () => {
-			done(rep.value);
+			resolve(JSON.parse(JSON.stringify(rep.value)));
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: 'foo',
 		object: {
 			numA: 1
@@ -80,12 +85,12 @@ test.serial('should accept the persisted value when it passes validation', async
 
 test.serial('should reject the persisted value when it fails validation, replacing with schemaDefaults', async t => {
 	// Persisted value is copied from fixtures
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaPersistenceFail');
-		rep.on('declared', () => done(rep.value));
-	});
+		rep.on('declared', () => resolve(JSON.parse(JSON.stringify(rep.value))));
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		object: {
 			numA: 0
@@ -94,7 +99,7 @@ test.serial('should reject the persisted value when it fails validation, replaci
 });
 
 test.serial('should accept valid assignment', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaAssignPass');
 		rep.once('declared', () => {
 			rep.value = {
@@ -106,13 +111,13 @@ test.serial('should accept valid assignment', async t => {
 
 			rep.on('change', newVal => {
 				if (newVal.string === 'foo') {
-					done(rep.value);
+					resolve(JSON.parse(JSON.stringify(rep.value)));
 				}
 			});
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: 'foo',
 		object: {
 			numA: 1
@@ -121,7 +126,7 @@ test.serial('should accept valid assignment', async t => {
 });
 
 test.serial('should throw on invalid assignment', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaAssignFail');
 		rep.once('declared', () => {
 			try {
@@ -129,28 +134,28 @@ test.serial('should throw on invalid assignment', async t => {
 					string: 0
 				};
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaAssignFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaAssignFail"'));
 });
 
 test.serial('should accept valid property deletion', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaDeletionPass');
 		rep.once('declared', () => {
 			delete rep.value.object.numB;
 			rep.on('change', newVal => {
 				if (!newVal.object.numB) {
-					done(rep.value);
+					resolve(JSON.parse(JSON.stringify(rep.value)));
 				}
 			});
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		object: {
 			numA: 0
@@ -159,34 +164,34 @@ test.serial('should accept valid property deletion', async t => {
 });
 
 test.serial('should throw on invalid property deletion', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaDeletionFail');
 		rep.once('declared', () => {
 			try {
 				delete rep.value.object.numA;
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaDeletionFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaDeletionFail"'));
 });
 
 test.serial('should accept valid array mutation via array mutator methods', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaArrayMutatorPass');
 		rep.once('declared', () => {
 			rep.value.array.push('foo');
 			rep.on('change', newVal => {
 				if (newVal.array.length === 1) {
-					done(rep.value);
+					resolve(JSON.parse(JSON.stringify(rep.value)));
 				}
 			});
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		array: [
 			'foo'
@@ -195,34 +200,34 @@ test.serial('should accept valid array mutation via array mutator methods', asyn
 });
 
 test.serial('should throw on invalid array mutation via array mutator methods', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaArrayMutatorFail');
 		rep.once('declared', () => {
 			try {
 				rep.value.array.push(0);
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaArrayMutatorFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaArrayMutatorFail"'));
 });
 
 test.serial('should accept valid property changes to arrays', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaArrayChangePass');
 		rep.once('declared', () => {
 			rep.value.array[0] = 'bar';
 			rep.on('change', newVal => {
 				if (newVal.array[0] === 'bar') {
-					done(rep.value);
+					resolve(JSON.parse(JSON.stringify(rep.value)));
 				}
 			});
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		array: [
 			'bar'
@@ -231,34 +236,34 @@ test.serial('should accept valid property changes to arrays', async t => {
 });
 
 test.serial('should throw on invalid property changes to arrays', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaArrayChangeFail');
 		rep.once('declared', () => {
 			try {
 				rep.value.array[0] = 0;
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaArrayChangeFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaArrayChangeFail"'));
 });
 
 test.serial('should accept valid property changes to objects', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaObjectChangePass');
 		rep.once('declared', () => {
 			rep.value.object.numA = 1;
 			rep.on('change', newVal => {
 				if (newVal.object.numA === 1) {
-					done(rep.value);
+					resolve(JSON.parse(JSON.stringify(rep.value)));
 				}
 			});
 		});
-	});
+	}));
 
-	t.deepEqual(res.value, {
+	t.deepEqual(res, {
 		string: '',
 		object: {
 			numA: 1
@@ -267,22 +272,22 @@ test.serial('should accept valid property changes to objects', async t => {
 });
 
 test.serial('should throw on invalid property changes to objects', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaObjectChangeFail');
 		rep.once('declared', () => {
 			try {
 				rep.value.object.numA = 'foo';
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Invalid value rejected for replicant "client_schemaObjectChangeFail"'));
+	t.true(res.startsWith('Invalid value rejected for replicant "client_schemaObjectChangeFail"'));
 });
 
 test.serial('should reject assignment if it was validated against a different version of the schema', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaAssignMismatch');
 		rep.once('declared', () => {
 			rep.schemaSum = 'baz';
@@ -294,36 +299,36 @@ test.serial('should reject assignment if it was validated against a different ve
 					}
 				};
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
 
 		rep.once('assignmentRejected', reason => {
-			done(reason);
+			resolve(reason);
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Mismatched schema version, assignment rejected'));
+	t.true(res.startsWith('Mismatched schema version, assignment rejected'));
 });
 
 test.serial('should reject operations if they were validated against a different version of the schema', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('client_schemaOperationMismatch');
 		rep.once('declared', () => {
 			rep.schemaSum = 'baz';
 			try {
 				rep.value.object.numA = 1;
 			} catch (e) {
-				done(e.message);
+				resolve(e.message);
 			}
 		});
 
 		rep.once('operationsRejected', reason => {
-			done(reason);
+			resolve(reason);
 		});
-	});
+	}));
 
-	t.true(res.value.startsWith('Mismatched schema version, assignment rejected'));
+	t.true(res.startsWith('Mismatched schema version, assignment rejected'));
 });
 
 // This isn't _actually_ testing schemas, it's just a bug that is easiest to detect when using a schema.
@@ -333,26 +338,32 @@ test.serial('should reject operations if they were validated against a different
 // This test should probably be re-written to be more targeted and remove any cruft.
 // Lange - 2017/05/04
 test.serial('shouldn\'t fuck up', async t => {
-	const res = await e.browser.client.executeAsync(done => {
+	const res = await dashboard.evaluate(() => new Promise(resolve => {
 		const rep = window.dashboardApi.Replicant('schedule:state');
 		rep.once('declared', () => {
 			rep.value.matchMap = [
-				false, false, false, false,
-				false, false, false, false
+				false,
+				false,
+				false,
+				false,
+				false,
+				false,
+				false,
+				false
 			];
 
 			rep.on('change', newVal => {
 				if (!newVal.matchMap.includes(true)) {
 					try {
 						rep.value.matchMap[6] = true;
-						done(true);
+						resolve(true);
 					} catch (e) {
-						done(e.message);
+						resolve(e.message);
 					}
 				}
 			});
 		});
-	});
+	}));
 
-	t.true(res.value);
+	t.true(res);
 });
