@@ -1,15 +1,18 @@
 'use strict';
 
+// Native
+import * as path from 'path';
+
 // Packages
-const path = require('path');
-const test = require('ava');
+import test from 'ava';
 
 // Ours
-require('../helpers/nodecg-and-webdriver')(test); // Must be first.
-const e = require('../helpers/test-environment');
+import * as server from '../helpers/server';
+
+server.setup();
 
 test('should create a default value based on the schema, if none is provided', t => {
-	const rep = e.apis.extension.Replicant('schema1');
+	const rep = t.context.apis.extension.Replicant('schema1');
 	t.deepEqual(rep.value, {
 		string: '',
 		object: {
@@ -20,7 +23,7 @@ test('should create a default value based on the schema, if none is provided', t
 
 test('should accept the defaultValue when it passes validation', t => {
 	t.notThrows(() => {
-		e.apis.extension.Replicant('schema2', {
+		t.context.apis.extension.Replicant('schema2', {
 			defaultValue: {
 				string: 'foo',
 				object: {
@@ -33,7 +36,7 @@ test('should accept the defaultValue when it passes validation', t => {
 
 test('should throw when defaultValue fails validation', t => {
 	const error = t.throws(() => {
-		e.apis.extension.Replicant('schema3', {
+		t.context.apis.extension.Replicant('schema3', {
 			defaultValue: {
 				string: 0
 			}
@@ -45,7 +48,7 @@ test('should throw when defaultValue fails validation', t => {
 
 test('should accept the persisted value when it passes validation', t => {
 	// Persisted value is copied from fixtures
-	const rep = e.apis.extension.Replicant('schemaPersistencePass');
+	const rep = t.context.apis.extension.Replicant('schemaPersistencePass');
 	t.deepEqual(rep.value, {
 		string: 'foo',
 		object: {
@@ -56,7 +59,7 @@ test('should accept the persisted value when it passes validation', t => {
 
 test('should reject the persisted value when it fails validation, replacing with schemaDefaults', t => {
 	// Persisted value is copied from fixtures
-	const rep = e.apis.extension.Replicant('schemaPersistenceFail');
+	const rep = t.context.apis.extension.Replicant('schemaPersistenceFail');
 	t.deepEqual(rep.value, {
 		string: '',
 		object: {
@@ -67,7 +70,7 @@ test('should reject the persisted value when it fails validation, replacing with
 
 test('should accept valid assignment', t => {
 	t.notThrows(() => {
-		const rep = e.apis.extension.Replicant('schemaAssignPass');
+		const rep = t.context.apis.extension.Replicant('schemaAssignPass');
 		rep.value = {
 			string: 'foo',
 			object: {
@@ -79,7 +82,7 @@ test('should accept valid assignment', t => {
 
 test('should throw on invalid assignment', t => {
 	const error = t.throws(() => {
-		const rep = e.apis.extension.Replicant('schemaAssignFail');
+		const rep = t.context.apis.extension.Replicant('schemaAssignFail');
 		rep.value = {
 			string: 0
 		};
@@ -90,14 +93,14 @@ test('should throw on invalid assignment', t => {
 
 test('should accept valid property deletion', t => {
 	t.notThrows(() => {
-		const rep = e.apis.extension.Replicant('schemaDeletionPass');
+		const rep = t.context.apis.extension.Replicant('schemaDeletionPass');
 		delete rep.value.object.numB;
 	});
 });
 
 test('should throw on invalid property deletion', t => {
 	const error = t.throws(() => {
-		const rep = e.apis.extension.Replicant('schemaDeletionFail');
+		const rep = t.context.apis.extension.Replicant('schemaDeletionFail');
 		delete rep.value.object.numA;
 	});
 
@@ -106,14 +109,14 @@ test('should throw on invalid property deletion', t => {
 
 test('should accept valid array mutation via array mutator methods', t => {
 	t.notThrows(() => {
-		const rep = e.apis.extension.Replicant('schemaArrayMutatorPass');
+		const rep = t.context.apis.extension.Replicant('schemaArrayMutatorPass');
 		rep.value.array.push('foo');
 	});
 });
 
 test('should throw on invalid array mutation via array mutator methods', t => {
 	const error = t.throws(() => {
-		const rep = e.apis.extension.Replicant('schemaArrayMutatorFail');
+		const rep = t.context.apis.extension.Replicant('schemaArrayMutatorFail');
 		rep.value.array.push(0);
 	});
 
@@ -122,14 +125,14 @@ test('should throw on invalid array mutation via array mutator methods', t => {
 
 test('should accept valid property changes to arrays', t => {
 	t.notThrows(() => {
-		const rep = e.apis.extension.Replicant('schemaArrayChangePass');
+		const rep = t.context.apis.extension.Replicant('schemaArrayChangePass');
 		rep.value.array[0] = 'bar';
 	});
 });
 
 test('should throw on invalid property changes to arrays', t => {
 	const error = t.throws(() => {
-		const rep = e.apis.extension.Replicant('schemaArrayChangeFail');
+		const rep = t.context.apis.extension.Replicant('schemaArrayChangeFail');
 		rep.value.array[0] = 0;
 	});
 
@@ -138,14 +141,14 @@ test('should throw on invalid property changes to arrays', t => {
 
 test('should accept valid property changes to objects', t => {
 	t.notThrows(() => {
-		const rep = e.apis.extension.Replicant('schemaObjectChangePass');
+		const rep = t.context.apis.extension.Replicant('schemaObjectChangePass');
 		rep.value.object.numA = 1;
 	});
 });
 
 test('should throw on invalid property changes to objects', t => {
 	const error = t.throws(() => {
-		const rep = e.apis.extension.Replicant('schemaObjectChangeFail');
+		const rep = t.context.apis.extension.Replicant('schemaObjectChangeFail');
 		rep.value.object.numA = 'foo';
 	});
 
@@ -153,7 +156,7 @@ test('should throw on invalid property changes to objects', t => {
 });
 
 test('should properly load schemas provided with an absolute path', t => {
-	const rep = e.apis.extension.Replicant('schemaAbsolutePath', {
+	const rep = t.context.apis.extension.Replicant('schemaAbsolutePath', {
 		schemaPath: path.resolve(__dirname, '../fixtures/nodecg-core/absolute-path-schemas/schemaAbsolutePath.json')
 	});
 
@@ -166,7 +169,7 @@ test('should properly load schemas provided with an absolute path', t => {
 });
 
 test('supports local file $refs', t => {
-	const rep = e.apis.extension.Replicant('schemaWithRef');
+	const rep = t.context.apis.extension.Replicant('schemaWithRef');
 	t.deepEqual(rep.schema, {
 		$schema: 'http://json-schema.org/draft-04/schema#',
 		type: 'object',
@@ -191,6 +194,10 @@ test('supports local file $refs', t => {
 								default: 0
 							}
 						}
+					},
+					hasFileRefThenDefRef: {
+						type: 'string',
+						enum: ['foo', 'bar']
 					}
 				}
 			}
@@ -199,7 +206,7 @@ test('supports local file $refs', t => {
 });
 
 test('supports internal $refs', t => {
-	const rep = e.apis.extension.Replicant('schemaWithInternalRef');
+	const rep = t.context.apis.extension.Replicant('schemaWithInternalRef');
 	t.deepEqual(rep.schema, {
 		$schema: 'http://json-schema.org/draft-04/schema#',
 		definitions: {
