@@ -3,8 +3,23 @@
 process.title = 'NodeCG';
 global.exitOnUncaught = true;
 
+const path = require('path');
 const cwd = process.cwd();
-global.isZeitPkg = __dirname.startsWith('/snapshot/') || __dirname.toLowerCase().startsWith('c:\\snapshot\\');
+global.isZeitPkg = isZeitPkg();
+
+/**
+ * Lange: As of July 2019, the only way I know to programmatically
+ * determine if code is running inside a Zeit pkg is to check if
+ * the first directory in the __dirname is called `snapshot`.
+ *
+ * There's definitely a possibility for false positives here,
+ * but I have yet to find a better alternative.
+ */
+function isZeitPkg() {
+	const parts = __dirname.split(path.sep);
+	return parts[1].toLowerCase() === 'snapshot';
+}
+
 if (global.isZeitPkg) {
 	console.info('[nodecg] Detected that NodeCG is running inside a ZEIT pkg (https://github.com/zeit/pkg)');
 } else if (cwd !== __dirname) {
@@ -18,9 +33,12 @@ if (!process.env.NODECG_ROOT) {
 }
 
 const semver = require('semver');
+const exitHook = require('exit-hook');
+const {engines} = require('./package.json');
 const nodeVersion = process.versions.node;
-if (!semver.satisfies(nodeVersion, '>=6')) {
-	console.error('ERROR: NodeCG requires Node.js >=6 and npm >=2');
+
+if (!semver.satisfies(nodeVersion, engines.node)) {
+	console.error(`ERROR: NodeCG requires Node.js ${engines.node}`);
 	console.error(`       Your Node.js version: v${nodeVersion}`);
 	process.exit(1);
 }
@@ -51,5 +69,9 @@ process.on('unhandledRejection', err => {
 const server = require('./lib/server')
 	.on('error', () => process.exit(1))
 	.on('stopped', () => process.exit(0));
+
+exitHook(() => {
+	server.stop();
+});
 
 server.start();
