@@ -4,12 +4,12 @@
 
 const Replicant = require('./replicant');
 const server = require('./server');
-const filteredConfig = require('./config').filteredConfig;
+const {filteredConfig} = require('./config');
 const utils = require('./util');
 const replicator = require('./replicator');
 const isError = require('is-error');
 const serializeError = require('serialize-error');
-const version = require('../package.json').version;
+const {version} = require('../package.json');
 let io;
 
 const apiContexts = new Set();
@@ -39,7 +39,7 @@ function NodeCG(bundle, socket) {
 
 	/**
 	 * The version (from package.json) of the bundle which this NodeCG API instance is for.
-	 * @name NodeCG#bundleName
+	 * @name NodeCG#bundleVersion
 	 */
 	this.bundleVersion = bundle.version;
 
@@ -101,7 +101,7 @@ function NodeCG(bundle, socket) {
 			const customCuesRep = new Replicant('customSoundCues', this.bundleName, {}, socket);
 			this._soundFiles = new Replicant('assets:sounds', this.bundleName, {}, socket);
 			this._bundleVolume = new Replicant(`volume:${this.bundleName}`, '_sounds', {}, socket);
-			this._masterVolume = new Replicant(`volume:master`, '_sounds', {}, socket);
+			this._masterVolume = new Replicant('volume:master', '_sounds', {}, socket);
 
 			this._soundCues = [];
 
@@ -350,9 +350,8 @@ NodeCG.sendMessageToBundle = function (messageName, bundleName, data, cb) {
 NodeCG.prototype.sendMessageToBundle = function (messageName, bundleName, data, cb) {
 	this.log.trace('Sending message %s to bundle %s with data:', messageName, bundleName, data);
 
-	/* eslint-disable prefer-spread */
+	// eslint-disable-next-line prefer-spread, prefer-rest-params
 	return NodeCG.sendMessageToBundle.apply(NodeCG, arguments);
-	/* eslint-enable prefer-spread */
 };
 /* eslint-enable no-unused-vars */
 
@@ -481,6 +480,7 @@ NodeCG.Replicant = function (name, namespace, opts) {
  * applied if this Replicant has not previously been declared and if it has no persisted value.
  * @param {boolean} [opts.persistent=true] - Whether to persist the Replicant's value to disk on every change.
  * Persisted values are re-loaded on startup.
+ * @param {number} [opts.persistenceInterval=DEFAULT_PERSISTENCE_INTERVAL] - Interval between each persistence, in milliseconds.
  * @param {string} [opts.schemaPath] - The filepath at which to look for a JSON Schema for this Replicant.
  * Defaults to `nodecg/bundles/${bundleName}/schemas/${replicantName}.json`. Please note that this default
  * path will be URIEncoded to ensure that it results in a valid filename.
@@ -805,7 +805,9 @@ function _registerSounds(ctx) {
 }
 
 function _setInstanceVolume(ctx, instance, cue) {
-	instance.volume = (ctx._masterVolume.value / 100) * (ctx._bundleVolume.value / 100) * (cue.volume / 100);
+	const volume = (ctx._masterVolume.value / 100) * (ctx._bundleVolume.value / 100) * (cue.volume / 100);
+	// Volue value must be finite or SoundJS throws error
+	instance.volume = isFinite(volume) ? volume : 0;
 }
 
 function _updateInstanceVolumes(ctx) {
@@ -835,6 +837,7 @@ function _wrapAcknowledgement(ack) {
 		if (handled) {
 			throw new Error('Acknowledgement already handled');
 		}
+
 		handled = true;
 
 		if (isError(firstArg)) {
