@@ -12,9 +12,8 @@ const sha1File = require('sha1-file');
 
 // Ours
 const AssetFile = require('./AssetFile');
-const authCheck = require('../util').authCheck;
+const {authCheck, debounceName} = require('../util');
 const bundles = require('../bundle-manager');
-const debounceName = require('../util').debounceName;
 const log = require('../logger')('nodecg/lib/assets');
 const Replicant = require('../replicant');
 
@@ -87,7 +86,11 @@ collections.forEach(({name, categories}) => {
 	});
 });
 
-const watcher = chokidar.watch(watchPatterns, {ignored: /[/\\]\./});
+// Chokidar no longer accepts Windows-style path separators when using globs.
+// Therefore, we must replace them with Unix-style ones.
+// See https://github.com/paulmillr/chokidar/issues/777 for more details.
+const fixedPaths = watchPatterns.map(pattern => pattern.replace(/\\/g, '/'));
+const watcher = chokidar.watch(fixedPaths, {ignored: /[/\\]\./});
 
 /* When the Chokidar watcher first starts up, it will fire an 'add' event for each file found.
  * After that, it will emit the 'ready' event.
@@ -176,6 +179,7 @@ app.get(
 				if (err.code === 'ENOENT') {
 					return res.sendStatus(404);
 				}
+
 				log.error(`Unexpected error sending file ${fullPath}`, err);
 				res.sendStatus(500);
 			}
@@ -224,9 +228,7 @@ app.delete(
 
 	// Delete the file (or an send appropriate error).
 	(req, res) => {
-		const namespace = req.params.namespace;
-		const category = req.params.category;
-		const filename = req.params.filename;
+		const {namespace, category, filename} = req.params;
 		const fullPath = path.join(ASSETS_ROOT, namespace, category, filename);
 
 		fs.unlink(fullPath, err => {
@@ -252,6 +254,7 @@ function calcNamespacedAssetsPath(namespace) {
 	if (!fs.existsSync(assetsPath)) {
 		fs.mkdirSync(assetsPath);
 	}
+
 	return assetsPath;
 }
 
