@@ -38,7 +38,7 @@ module.exports = function (userCfgPath) {
 			default: ''
 		},
 		developer: {
-			doc: 'Whether to enable features that speed up development. Not suitable for production.',
+			doc: 'Deprecated, currently does nothing.',
 			format: Boolean,
 			default: false
 		},
@@ -91,6 +91,14 @@ module.exports = function (userCfgPath) {
 				},
 				default: null,
 				arg: 'bundlesDisabled'
+			},
+			paths: {
+				doc: 'An array of additional paths where bundles are located',
+				format(val) {
+					return VALIDATE_STRING_ARRAY(val); // eslint-disable-line new-cap
+				},
+				default: [],
+				arg: 'bundlesPaths'
 			}
 		}
 	};
@@ -225,6 +233,11 @@ module.exports = function (userCfgPath) {
 
 				// This will throw if the user does not provide a value, but only if SSL is enabled.
 				default: userConfig.ssl && userConfig.ssl.enabled ? null : ''
+			},
+			passphrase: {
+				doc: 'The passphrase for the provided key file.',
+				format: String,
+				default: ''
 			}
 		};
 
@@ -269,10 +282,6 @@ module.exports = function (userCfgPath) {
 
 	if (!config.sentry) {
 		config.sentry = {};
-	}
-
-	if (config.developer) {
-		console.warn('[nodecg] Developer mode is active! Be sure to disable this in production.');
 	}
 
 	// Create the filtered config
