@@ -3,11 +3,9 @@
 // Native
 const fs = require('fs');
 const path = require('path');
-const pug = require('pug');
 
 // Packages
 const clone = require('clone');
-const debounce = require('lodash.debounce');
 const express = require('express');
 
 // Ours
@@ -21,14 +19,12 @@ const app = express();
 const SRC_PATH = path.join(__dirname, '../src');
 const INSTRUMENTED_PATH = path.join(__dirname, '../instrumented');
 const BUILD_PATH = path.join(__dirname, '../build/src');
-let cachedDashboard = null;
-const recacheDashboard = debounce(() => {
-	cachedDashboard = renderDashboard();
-}, 100);
+
+let dashboardContext = null;
 
 log.trace('Adding Express routes');
 
-app.use('/bower_components', express.static(path.resolve(__dirname, '../bower_components')));
+app.use('/node_modules', express.static(path.resolve(__dirname, '../node_modules')));
 
 app.get('/', (req, res) => res.redirect('/dashboard/'));
 
@@ -37,53 +33,47 @@ app.get('/dashboard', ncgUtils.authCheck, (req, res) => {
 		return res.redirect('/dashboard/');
 	}
 
-	if (configHelper.config.developer) {
-		res.send(renderDashboard());
-	} else {
-		res.status(200).send(cachedDashboard);
+	if (!dashboardContext) {
+		dashboardContext = getDashboardContext();
 	}
+
+	res.render(path.join(__dirname, '../src/dashboard/dashboard.tmpl'), dashboardContext);
 });
 
-if (process.env.NODECG_TEST || configHelper.config.developer) {
-	app.get('/nodecg-api.min.js', (req, res) => {
-		res.sendFile(path.join(process.env.NODECG_TEST ? INSTRUMENTED_PATH : BUILD_PATH, 'nodecg-api.min.js'));
-	});
+app.get('/nodecg-api.min.js', (req, res) => {
+	res.sendFile(path.join(process.env.NODECG_TEST ? INSTRUMENTED_PATH : BUILD_PATH, 'nodecg-api.min.js'));
+});
 
-	app.get('/nodecg-api.min.js.map', (req, res) => {
-		res.sendFile(path.join(process.env.NODECG_TEST ? INSTRUMENTED_PATH : BUILD_PATH, 'nodecg-api.min.js.map'));
-	});
+app.get('/nodecg-api.min.js.map', (req, res) => {
+	res.sendFile(path.join(process.env.NODECG_TEST ? INSTRUMENTED_PATH : BUILD_PATH, 'nodecg-api.min.js.map'));
+});
 
-	if (process.env.NODECG_TEST) {
-		log.warn('Serving instrumented (and also unbundled) files for testing');
-		app.get('/*', (req, res, next) => {
-			const resName = req.params[0];
-			if (!resName.startsWith('dashboard/') && !resName.startsWith('instance/')) {
-				return next();
-			}
-
-			const fp = path.join(INSTRUMENTED_PATH, resName);
-			if (fs.existsSync(fp)) {
-				return res.sendFile(fp, err => {
-					/* istanbul ignore next */
-					if (err && !res.headersSent) {
-						return next();
-					}
-				});
-			}
-
+if (process.env.NODECG_TEST) {
+	log.warn('Serving instrumented files for testing');
+	app.get('/*', (req, res, next) => {
+		const resName = req.params[0];
+		if (!resName.startsWith('dashboard/') && !resName.startsWith('instance/')) {
 			return next();
-		});
-	} else {
-		log.warn('Serving unbundled files for development');
-	}
+		}
 
-	app.use('/', express.static(SRC_PATH));
-} else {
-	app.use('/', express.static(BUILD_PATH));
+		const fp = path.join(INSTRUMENTED_PATH, resName);
+		if (fs.existsSync(fp)) {
+			return res.sendFile(fp, err => {
+				/* istanbul ignore next */
+				if (err && !res.headersSent) {
+					return next();
+				}
+			});
+		}
+
+		return next();
+	});
 }
 
+app.use('/', express.static(SRC_PATH));
+
 app.get('/bundles/:bundleName/dashboard/*', ncgUtils.authCheck, (req, res, next) => {
-	const bundleName = req.params.bundleName;
+	const {bundleName} = req.params;
 	const bundle = bundles.find(bundleName);
 	if (!bundle) {
 		next();
@@ -106,7 +96,7 @@ app.get('/bundles/:bundleName/dashboard/*', ncgUtils.authCheck, (req, res, next)
 		res.sendFile(fileLocation, err => {
 			if (err) {
 				if (err.code === 'ENOENT') {
-					return res.sendStatus(404);
+					return next();
 				}
 
 				/* istanbul ignore next */
@@ -118,15 +108,15 @@ app.get('/bundles/:bundleName/dashboard/*', ncgUtils.authCheck, (req, res, next)
 	}
 });
 
-// Initialize cache
-recacheDashboard();
+// When a bundle changes, delete the cached dashboard context
+bundles.on('bundleChanged', () => {
+	dashboardContext = null;
+});
 
-// When a bundle changes, re-cache the dashboard.
-bundles.on('bundleChanged', recacheDashboard);
 module.exports = app;
 
-function renderDashboard() {
-	return pug.renderFile(path.join(__dirname, '../src/dashboard/dashboard.pug'), {
+function getDashboardContext() {
+	return {
 		bundles: bundles.all().map(bundle => {
 			const cleanedBundle = clone(bundle);
 			if (cleanedBundle.dashboardPanels) {
@@ -142,7 +132,7 @@ function renderDashboard() {
 		privateConfig: configHelper.config,
 		workspaces: parseWorkspaces(),
 		ravenConfig
-	});
+	};
 }
 
 function parseWorkspaces() {
@@ -181,14 +171,6 @@ function parseWorkspaces() {
 		});
 	});
 
-	if (defaultWorkspaceHasPanels || !otherWorkspacesHavePanels) {
-		workspaces.unshift({
-			name: 'default',
-			label: otherWorkspacesHavePanels ? 'Main Workspace' : 'Workspace',
-			route: ''
-		});
-	}
-
 	workspaceNames.forEach(name => {
 		workspaces.push({
 			name,
@@ -196,6 +178,18 @@ function parseWorkspaces() {
 			route: `workspace/${name}`
 		});
 	});
+
+	workspaces.sort((a, b) => {
+		return a.label.localeCompare(b.label);
+	});
+
+	if (defaultWorkspaceHasPanels || !otherWorkspacesHavePanels) {
+		workspaces.unshift({
+			name: 'default',
+			label: otherWorkspacesHavePanels ? 'Main Workspace' : 'Workspace',
+			route: ''
+		});
+	}
 
 	return workspaces;
 }
