@@ -19,7 +19,7 @@ app.set('views', path.resolve(__dirname, '..'));
 app.use(require('./registration'));
 
 app.get('/bundles/:bundleName/graphics*', ncgUtils.authCheck, (req, res, next) => {
-	const bundleName = req.params.bundleName;
+	const {bundleName} = req.params;
 	const bundle = Bundles.find(bundleName);
 	if (!bundle) {
 		next();
@@ -68,7 +68,7 @@ app.get('/bundles/:bundleName/graphics*', ncgUtils.authCheck, (req, res, next) =
 		res.sendFile(fileLocation, err => {
 			if (err) {
 				if (err.code === 'ENOENT') {
-					return res.sendStatus(404);
+					return next();
 				}
 
 				/* istanbul ignore next */
@@ -82,7 +82,7 @@ app.get('/bundles/:bundleName/graphics*', ncgUtils.authCheck, (req, res, next) =
 
 // This isn't really a graphics-specific thing, should probably be in the main server lib.
 app.get('/bundles/:bundleName/:target(bower_components|node_modules)/*', (req, res, next) => {
-	const bundleName = req.params.bundleName;
+	const {bundleName} = req.params;
 	const bundle = Bundles.find(bundleName);
 	if (!bundle) {
 		next();
