@@ -8,11 +8,11 @@ const app = require('express')();
 
 // Ours
 const bundles = require('../bundle-manager');
-const injectScripts = require('../util').injectScripts;
+const {injectScripts} = require('../util');
 const io = require('../server').getIO();
 const Replicant = require('../replicant');
 
-const BUILD_PATH = path.join(__dirname, '../../build/src/instance');
+const BUILD_PATH = path.join(__dirname, '../../src/instance');
 const instancesRep = new Replicant('graphics:instances', 'nodecg', {
 	schemaPath: path.resolve(__dirname, '../../schemas/graphics%3Ainstances.json'),
 	persistent: false
@@ -23,8 +23,8 @@ bundles.on('gitChanged', updateInstanceStatuses);
 
 io.on('connection', socket => {
 	socket.on('graphic:registerSocket', (regRequest, cb) => {
-		const bundleName = regRequest.bundleName;
-		let pathName = regRequest.pathName;
+		const {bundleName} = regRequest;
+		let {pathName} = regRequest;
 		if (pathName.endsWith(`/${bundleName}/graphics/`)) {
 			pathName += 'index.html';
 		}
@@ -211,8 +211,7 @@ function calcBundleVersionMismatch(bundle, regRequest) {
 
 function updateInstanceStatuses() {
 	instancesRep.value.forEach(instance => {
-		const bundleName = instance.bundleName;
-		const pathName = instance.pathName;
+		const {bundleName, pathName} = instance;
 		const bundle = bundles.find(bundleName);
 		/* istanbul ignore next: simple error trapping */
 		if (!bundle) {
