@@ -16,7 +16,7 @@ const LOG_LEVELS = {
  * Enum logging level values.
  * @enum {String}
  */
-const ENUM_LEVELS = { // eslint-disable-line no-unused-vars
+const ENUM_LEVELS = {// eslint-disable-line no-unused-vars
 	trace: 'The highest level of logging, logs everything.',
 	debug: 'Less spammy than trace, includes most info relevant for debugging.',
 	info: 'The default logging level. Logs useful info, warnings, and errors.',
@@ -53,7 +53,7 @@ module.exports = function (initialOpts, Raven) {
 			this.name = name;
 		}
 
-		trace() {
+		trace(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -62,10 +62,10 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.info(`[${this.name}]`, ...arguments);
+			console.info(`[${this.name}]`, ...args);
 		}
 
-		debug() {
+		debug(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -74,10 +74,10 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.info(`[${this.name}]`, ...arguments);
+			console.info(`[${this.name}]`, ...args);
 		}
 
-		info() {
+		info(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -86,10 +86,10 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.info(`[${this.name}]`, ...arguments);
+			console.info(`[${this.name}]`, ...args);
 		}
 
-		warn() {
+		warn(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -98,10 +98,10 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.warn(`[${this.name}]`, ...arguments);
+			console.warn(`[${this.name}]`, ...args);
 		}
 
-		error() {
+		error(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -110,10 +110,10 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.error(`[${this.name}]`, ...arguments);
+			console.error(`[${this.name}]`, ...args);
 
 			if (Raven) {
-				const formattedArgs = Array.from(arguments).map(argument => {
+				const formattedArgs = args.map(argument => {
 					return typeof argument === 'object' ?
 						inspect(argument, {depth: null, showProxy: true}) :
 						argument;
@@ -125,7 +125,7 @@ module.exports = function (initialOpts, Raven) {
 			}
 		}
 
-		replicants() {
+		replicants(...args) {
 			if (Logger._silent) {
 				return;
 			}
@@ -134,7 +134,7 @@ module.exports = function (initialOpts, Raven) {
 				return;
 			}
 
-			console.info(`[${this.name}]`, ...arguments);
+			console.info(`[${this.name}]`, ...args);
 		}
 
 		static globalReconfigure(opts) {
