@@ -9,7 +9,7 @@ const winston = require('winston');
  * Enum logging level values.
  * @enum {String}
  */
-const ENUM_LEVELS = { // eslint-disable-line no-unused-vars
+const ENUM_LEVELS = {// eslint-disable-line no-unused-vars
 	trace: 'The highest level of logging, logs everything.',
 	debug: 'Less spammy than trace, includes most info relevant for debugging.',
 	info: 'The default logging level. Logs useful info, warnings, and errors.',
@@ -90,32 +90,27 @@ module.exports = function (initialOpts, Raven) {
 			this.name = name;
 		}
 
-		trace() {
-			const args = Array.from(arguments);
+		trace(...args) {
 			mainLogger.trace(`[${this.name}] ${args[0]}`, ...args.slice(1));
 		}
 
-		debug() {
-			const args = Array.from(arguments);
+		debug(...args) {
 			mainLogger.debug(`[${this.name}] ${args[0]}`, ...args.slice(1));
 		}
 
-		info() {
-			const args = Array.from(arguments);
+		info(...args) {
 			mainLogger.info(`[${this.name}] ${args[0]}`, ...args.slice(1));
 		}
 
-		warn() {
-			const args = Array.from(arguments);
+		warn(...args) {
 			mainLogger.warn(`[${this.name}] ${args[0]}`, ...args.slice(1));
 		}
 
-		error() {
-			const args = Array.from(arguments);
+		error(...args) {
 			mainLogger.error(`[${this.name}] ${args[0]}`, ...args.slice(1));
 
 			if (Raven) {
-				const formattedArgs = Array.from(arguments).map(argument => {
+				const formattedArgs = args.map(argument => {
 					return typeof argument === 'object' ?
 						inspect(argument, {depth: null, showProxy: true}) :
 						argument;
@@ -127,12 +122,11 @@ module.exports = function (initialOpts, Raven) {
 			}
 		}
 
-		replicants() {
+		replicants(...args) {
 			if (!Logger._shouldLogReplicants) {
 				return;
 			}
 
-			const args = Array.from(arguments);
 			mainLogger.info(`[${this.name}] ${args[0]}`, ...args.slice(1));
 		}
 
