@@ -3,7 +3,7 @@
 const path = require('path');
 const express = require('express');
 const app = express();
-const config = require('../config').config;
+const {config} = require('../config');
 const cookieParser = require('cookie-parser');
 const session = require('express-session');
 const NedbStore = require('express-nedb-session')(session);
@@ -42,6 +42,7 @@ if (config.login.twitch && config.login.twitch.enabled) {
 	if (scope.indexOf('user:read:email') < 0) {
 		scope.push('user:read:email');
 	}
+
 	scope = scope.join(' ');
 
 	passport.use(new TwitchStrategy({
@@ -55,7 +56,7 @@ if (config.login.twitch && config.login.twitch.enabled) {
 		if (profile.allowed) {
 			log.info('Granting %s access', profile.username);
 			profile.accessToken = accessToken;
-			// Twitch oauth does not use refreshToken
+			profile.refreshToken = refreshToken;
 		} else {
 			log.info('Denying %s access', profile.username);
 		}
@@ -65,22 +66,45 @@ if (config.login.twitch && config.login.twitch.enabled) {
 }
 
 if (config.login.local && config.login.local.enabled) {
-	const LocalStrategy = require('passport-local').Strategy;
+	const {Strategy: LocalStrategy} = require('passport-local');
+	const crypto = require('crypto');
+
+	const {sessionSecret, local: {allowedUsers}} = config.login;
+	const hashes = crypto.getHashes();
 
 	passport.use(new LocalStrategy({
 		usernameField: 'username',
 		passwordField: 'password',
 		session: false
-	}, (user, password, done) => {
-		const isAllowed = (config.login.local.allowedUsers.findIndex(obj => obj.username === user && obj.password === password) > -1);
-		const profile = {username: user, allowed: isAllowed, provider: 'local'};
-		if (profile.allowed) {
-			log.info('Granting %s access using local auth', profile.username);
-		} else {
-			log.info('Denying %s access using local auth', profile.username);
+	}, (username, password, done) => {
+		const user = allowedUsers.find(u => u.username === username);
+		let allowed = false;
+
+		if (user) {
+			const match = user.password.match(/^([^:]+):(.+)$/);
+			let expected = user.password;
+			let actual = password;
+
+			if (match && hashes.includes(match[1])) {
+				expected = match[2];
+				actual = crypto
+					.createHmac(match[1], sessionSecret)
+					.update(actual, 'utf8')
+					.digest('hex');
+			}
+
+			if (expected === actual) {
+				allowed = true;
+			}
 		}
 
-		return done(null, profile);
+		log.info('%s %s access using local auth', allowed ? 'Granting' : 'Denying', username);
+
+		return done(null, {
+			provider: 'local',
+			username,
+			allowed
+		});
 	}));
 }
 
@@ -105,14 +129,14 @@ app.use('/login', express.static(path.join(__dirname, 'public')));
 app.set('views', __dirname);
 
 app.get('/login', (req, res) => {
-	res.render('public/login.pug', {
+	res.render('public/login.tmpl', {
 		user: req.user,
 		config
 	});
 });
 
 app.get('/authError', (req, res) => {
-	res.render('public/authError.pug', {
+	res.render('public/authError.tmpl', {
 		message: req.query.message,
 		code: req.query.code,
 		viewUrl: req.query.viewUrl
