@@ -20,7 +20,7 @@ all.forEach(bundle => {
 			res.sendFile(fileLocation, err => {
 				if (err) {
 					if (err.code === 'ENOENT') {
-						return res.sendStatus(404);
+						return next();
 					}
 
 					/* istanbul ignore next */
