@@ -3,12 +3,13 @@
 // Native
 const EventEmitter = require('events');
 const fs = require('fs');
+const throttle = require('lodash.throttle');
 const path = require('path');
 
 // Packages
 const $RefParser = require('json-schema-lib');
 const clone = require('clone');
-const LocalStorage = require('node-localstorage').LocalStorage;
+const {LocalStorage} = require('node-localstorage');
 const schemaDefaults = require('json-schema-defaults');
 const sha1 = require('sha1');
 
@@ -66,6 +67,10 @@ class Replicant extends EventEmitter {
 			opts.persistent = true;
 		}
 
+		if (typeof opts.persistenceInterval === 'undefined') {
+			opts.persistenceInterval = shared.DEFAULT_PERSISTENCE_INTERVAL;
+		}
+
 		this.name = name;
 		this.namespace = namespace;
 		this.opts = opts;
@@ -113,6 +118,11 @@ class Replicant extends EventEmitter {
 			existingValue = JSON.parse(existingValue);
 		} catch (e) {}
 
+		// Set the default value, if a schema is present and no default value was provided.
+		if (this.schema && typeof opts.defaultValue === 'undefined') {
+			opts.defaultValue = schemaDefaults(this.schema);
+		}
+
 		// If `opts.persistent` is true and this replicant has a persisted value, try to load that persisted value.
 		// Else, apply `opts.defaultValue`.
 		if (opts.persistent && typeof existingValue !== 'undefined' && existingValue !== null) {
@@ -125,9 +135,7 @@ class Replicant extends EventEmitter {
 			}
 		} else {
 			if (this.schema) {
-				if (typeof opts.defaultValue === 'undefined') {
-					opts.defaultValue = schemaDefaults(this.schema);
-				} else {
+				if (typeof opts.defaultValue !== 'undefined') {
 					this.validate(opts.defaultValue);
 				}
 			}
@@ -159,6 +167,11 @@ class Replicant extends EventEmitter {
 				listener(this.value);
 			}
 		});
+
+		this._requestSaveReplicant = throttle(
+			() => replicator.saveReplicant(this),
+			this.opts.persistenceInterval,
+		);
 
 		return this;
 	}
@@ -199,7 +212,7 @@ class Replicant extends EventEmitter {
 			operations: this._operationQueue,
 			revision: this.revision
 		});
-		replicator.saveReplicant(this);
+		this._requestSaveReplicant();
 		this.emit('change', this.value, this._oldValue, this._operationQueue);
 		this._operationQueue = [];
 	}
