@@ -25,7 +25,17 @@ function replaceRefs(obj, currentFile, allFiles) {
 			if (!referenceFile) {
 				throw new Error('Should have been a schema here but wasn\'t');
 			}
+
 			dereferencedData = referenceFile.data;
+
+			// If this file reference also has a local reference appended to it,
+			// we need to resolve that local reference within the file we just dereferenced.
+			// Example: schemaRefTargetWithDef.json#/definitions/exampleDef
+			const hashIndex = obj.$ref.indexOf('#');
+			if (hashIndex >= 0) {
+				const hashPath = obj.$ref.slice(hashIndex);
+				dereferencedData = resolvePointerReference(dereferencedData, hashPath);
+			}
 		} else if (isPointerReference(obj)) {
 			referenceFile = currentFile;
 			dereferencedData = resolvePointerReference(currentFile.data, obj.$ref);
@@ -37,6 +47,7 @@ function replaceRefs(obj, currentFile, allFiles) {
 				if (key === '$schema') {
 					continue;
 				}
+
 				obj[key] = clone(dereferencedData[key]);
 			}
 
@@ -90,7 +101,7 @@ function isPointerReference(value) {
  * @param {File} file - The file that the JSON Reference is in
  */
 function resolveFileReference(url, file) {
-	const schema = file.schema;
+	const {schema} = file;
 
 	// Remove any hash from the URL, since this URL represents the WHOLE file, not a fragment of it
 	url = jsonSchemaStripHash(url);
