@@ -20,9 +20,15 @@ const ARRAY_MUTATOR_METHODS = [
 	'unshift'
 ];
 
+/**
+ * The default persistence interval, in milliseconds.
+ * @type {Number}
+ */
+const DEFAULT_PERSISTENCE_INTERVAL = 100;
+
 const deleteTrap = function (target, prop) {
 	const metadata = metadataMap.get(target);
-	const replicant = metadata.replicant;
+	const {replicant} = metadata;
 
 	if (replicant._ignoreProxy) {
 		return delete target[prop];
@@ -49,7 +55,7 @@ const deleteTrap = function (target, prop) {
 const CHILD_ARRAY_HANDLER = {
 	get(target, prop) {
 		const metadata = metadataMap.get(target);
-		const replicant = metadata.replicant;
+		const {replicant} = metadata;
 		if (metadata.replicant._ignoreProxy) {
 			return target[prop];
 		}
@@ -57,20 +63,20 @@ const CHILD_ARRAY_HANDLER = {
 		if ({}.hasOwnProperty.call(Array.prototype, prop) && typeof Array.prototype[prop] === 'function' &&
 			target[prop] === Array.prototype[prop] && ARRAY_MUTATOR_METHODS.indexOf(prop) >= 0) {
 			/* eslint-disable prefer-spread */
-			return function () {
+			return (...args) => {
 				if (replicant.schema) {
 					const valueClone = clone(replicant.value);
 					const targetClone = objectPath.get(valueClone, pathStrToPathArr(metadata.path));
-					targetClone[prop].apply(targetClone, arguments);
+					targetClone[prop].apply(targetClone, args);
 					replicant.validate(valueClone);
 				}
 
 				if (process.browser) {
-					metadata.replicant._addOperation(metadata.path, prop, Array.prototype.slice.call(arguments));
+					metadata.replicant._addOperation(metadata.path, prop, Array.prototype.slice.call(args));
 				} else {
 					replicant._ignoreProxy = true;
-					metadata.replicant._addOperation(metadata.path, prop, Array.prototype.slice.call(arguments));
-					const retValue = target[prop].apply(target, arguments);
+					metadata.replicant._addOperation(metadata.path, prop, Array.prototype.slice.call(args));
+					const retValue = target[prop].apply(target, args);
 					replicant._ignoreProxy = false;
 
 					// We have to re-proxy the target because the items could have been inserted.
@@ -92,7 +98,7 @@ const CHILD_ARRAY_HANDLER = {
 		}
 
 		const metadata = metadataMap.get(target);
-		const replicant = metadata.replicant;
+		const {replicant} = metadata;
 
 		if (replicant._ignoreProxy) {
 			target[prop] = newValue;
@@ -154,7 +160,7 @@ const CHILD_OBJECT_HANDLER = {
 		}
 
 		const metadata = metadataMap.get(target);
-		const replicant = metadata.replicant;
+		const {replicant} = metadata;
 
 		if (replicant._ignoreProxy) {
 			target[prop] = newValue;
@@ -194,6 +200,7 @@ const CHILD_OBJECT_HANDLER = {
 
 module.exports = {
 	ARRAY_MUTATOR_METHODS,
+	DEFAULT_PERSISTENCE_INTERVAL,
 	_proxyRecursive: proxyRecursive,
 
 	/**
@@ -222,7 +229,7 @@ module.exports = {
 				case 'update': {
 					path.push(operation.args.prop);
 
-					let newValue = operation.args.newValue;
+					let {newValue} = operation.args;
 					if (typeof newValue === 'object') {
 						newValue = this._proxyRecursive(replicant, newValue, pathArrToPathStr(path));
 					}
@@ -230,12 +237,14 @@ module.exports = {
 					result = objectPath.set(replicant.value, path, newValue);
 					break;
 				}
+
 				case 'delete':
 					// Workaround for https://github.com/mariocasciaro/object-path/issues/69
 					if (path.length === 0 || objectPath.has(replicant.value, path)) {
 						const target = objectPath.get(replicant.value, path);
 						result = delete target[operation.args.prop];
 					}
+
 					break;
 				/* istanbul ignore next */
 				default:
