@@ -25,6 +25,7 @@ module.exports = {
 	emitToClients,
 	saveReplicant,
 	declaredReplicants,
+	saveAllReplicants,
 	stores
 };
 
@@ -260,7 +261,7 @@ function saveReplicant(replicant) {
 			return;
 		}
 
-		let value = replicant.value;
+		let {value} = replicant;
 		if (typeof value === 'object') {
 			value = JSON.stringify(value);
 		} else if (typeof value === 'undefined') {
@@ -271,7 +272,19 @@ function saveReplicant(replicant) {
 			stores[replicant.namespace].setItem(`${replicant.name}.rep`, value);
 			replicant.log.replicants('Value successfully persisted.');
 		} catch (error) {
+			if (error.name !== 'QUOTA_EXCEEDED_ERR') {
+				return replicant._requestSaveReplicant();
+			}
+
 			replicant.log.error('Failed to persist value:', error);
 		}
 	});
+}
+
+function saveAllReplicants() {
+	for (const replicants of Object.values(declaredReplicants)) {
+		for (const replicant of Object.values(replicants)) {
+			saveReplicant(replicant);
+		}
+	}
 }
