@@ -1,6 +1,6 @@
 'use strict';
 
-const EventEmitter = require('events').EventEmitter;
+const {EventEmitter} = require('events');
 const path = require('path');
 const semver = require('semver');
 const Raven = require('raven');
@@ -68,6 +68,7 @@ module.exports.init = function () {
 					if (!{}.hasOwnProperty.call(bundle.bundleDependencies, dep)) {
 						continue;
 					}
+
 					/* istanbul ignore if */
 					if (bundle._satisfiedDepNames.indexOf(dep) > -1) {
 						continue;
