@@ -2,7 +2,7 @@
 
 const pjson = require('../../package.json');
 const configHelper = require('../config');
-const config = configHelper.config;
+const {config} = configHelper;
 global.exitOnUncaught = config.exitOnUncaught;
 
 const Raven = require('raven');
@@ -20,7 +20,7 @@ if (config.sentry && config.sentry.enabled) {
 }
 
 // Native
-const EventEmitter = require('events').EventEmitter;
+const {EventEmitter} = require('events');
 const fs = require('fs');
 const path = require('path');
 
@@ -31,6 +31,9 @@ const debounce = require('lodash.debounce');
 const express = require('express');
 const fetch = require('make-fetch-happen');
 const semver = require('semver');
+const template = require('lodash.template');
+const memoize = require('fast-memoize');
+const transformMiddleware = require('express-transform-bare-module-specifiers').default;
 
 // Ours
 const bundleManager = require('../bundle-manager');
@@ -56,6 +59,10 @@ fetch('http://registry.npmjs.org/nodecg/latest').then(res => {
 	// Discard errors.
 });
 
+const renderTemplate = memoize((content, options) => {
+	return template(content)(options);
+});
+
 module.exports = new EventEmitter();
 
 module.exports.start = function () {
@@ -74,6 +81,9 @@ module.exports.start = function () {
 			key: fs.readFileSync(config.ssl.keyPath),
 			cert: fs.readFileSync(config.ssl.certificatePath)
 		};
+		if (config.ssl.passphrase) {
+			sslOpts.passphrase = config.ssl.passphrase;
+		}
 
 		// If we allow HTTP on the same port, use httpolyglot
 		// otherwise, standard https server
@@ -83,6 +93,7 @@ module.exports.start = function () {
 	} else {
 		server = require('http').createServer(app);
 	}
+
 	io = require('socket.io')(server);
 	io.sockets.setMaxListeners(64); // Prevent console warnings when many extensions are installed
 
@@ -91,6 +102,16 @@ module.exports.start = function () {
 	app.use(require('compression')());
 	app.use(bodyParser.json());
 	app.use(bodyParser.urlencoded({extended: true}));
+
+	app.engine('tmpl', (filePath, options, callback) => {
+		fs.readFile(filePath, (error, content) => {
+			if (error) {
+				return callback(error);
+			}
+
+			return callback(null, renderTemplate(content, options));
+		});
+	});
 
 	if (config.login && config.login.enabled) {
 		log.info('Login security enabled');
@@ -103,11 +124,37 @@ module.exports.start = function () {
 		});
 	}
 
-	const bundlesPath = global.isZeitPkg ?
-		path.resolve(__dirname, '../../bundles') :
-		path.join(process.env.NODECG_ROOT, 'bundles');
+	const bundlesPaths = [path.join(process.env.NODECG_ROOT, 'bundles')].concat(config.bundles.paths);
 	const cfgPath = path.join(process.env.NODECG_ROOT, 'cfg');
-	bundleManager.init(bundlesPath, cfgPath, require('../../package.json').version, config, Logger);
+
+	if (global.isZeitPkg) {
+		bundlesPaths.unshift(path.resolve(__dirname, '../../bundles'));
+	}
+
+	const dashboardTransformRootDir = path.resolve(__dirname, '../..');
+	app.use('/node_modules/*', transformMiddleware({
+		rootDir: path.join(dashboardTransformRootDir, 'node_modules'),
+		modulesUrl: '/node_modules'
+	}));
+
+	app.use('/dashboard/*', transformMiddleware({
+		rootDir: path.join(dashboardTransformRootDir, 'src'),
+		modulesUrl: '/node_modules'
+	}));
+
+	bundleManager.init(bundlesPaths, cfgPath, pjson.version, config, Logger);
+
+	bundleManager.all().forEach(bundle => {
+		if (bundle.transformBareModuleSpecifiers) {
+			const opts = {
+				rootDir: global.isZeitPkg ?
+					path.resolve(__dirname, '../../') :
+					process.env.NODECG_ROOT,
+				modulesUrl: `/bundles/${bundle.name}/node_modules`
+			};
+			app.use(`/bundles/${bundle.name}/*`, transformMiddleware(opts));
+		}
+	});
 
 	io.on('error', err => {
 		if (global.sentryEnabled) {
@@ -137,17 +184,19 @@ module.exports.start = function () {
 			if (typeof room !== 'string') {
 				throw new Error('Room must be a string');
 			}
+
 			if (Object.keys(socket.rooms).indexOf(room) < 0) {
 				log.trace('Socket %s joined room:', socket.id, room);
 				socket.join(room);
 			}
+
 			if (typeof cb === 'function') {
 				cb();
 			}
 		});
 
 		if (config.login && config.login.enabled) {
-			const token = socket.token;
+			const {token} = socket;
 			if (!{}.hasOwnProperty.call(authorizedSockets, token)) {
 				authorizedSockets[token] = [];
 			}
@@ -207,6 +256,7 @@ module.exports.start = function () {
 				if (process.env.NODECG_TEST) {
 					return;
 				}
+
 				log.error(`[server.js] Listen ${config.host}:${config.port} in use, is NodeCG already running? NodeCG will now exit.`);
 				break;
 			default:
@@ -269,7 +319,7 @@ module.exports.start = function () {
 		port: process.env.NODECG_TEST ? undefined : config.port
 	}, () => {
 		if (process.env.NODECG_TEST) {
-			const port = server.address().port;
+			const {port} = server.address();
 			log.warn(`Test mode active, using automatic listen port: ${port}`);
 			configHelper.config.port = port;
 			configHelper.filteredConfig.port = port;
@@ -290,6 +340,8 @@ module.exports.stop = function () {
 	if (io) {
 		io.close();
 	}
+
+	require('../replicator').saveAllReplicants();
 
 	extensionManager = null;
 	io = null;
