@@ -14,7 +14,7 @@ const app = express();
 log.trace('Adding Express routes');
 
 app.get('/bundles/:bundleName/shared/*', ncgUtils.authCheck, (req, res, next) => {
-	const bundleName = req.params.bundleName;
+	const {bundleName} = req.params;
 	const bundle = bundles.find(bundleName);
 	if (!bundle) {
 		next();
@@ -29,7 +29,7 @@ app.get('/bundles/:bundleName/shared/*', ncgUtils.authCheck, (req, res, next) =>
 	res.sendFile(fileLocation, err => {
 		if (err) {
 			if (err.code === 'ENOENT') {
-				return res.sendStatus(404);
+				return next();
 			}
 
 			/* istanbul ignore next */
