@@ -1,7 +1,7 @@
 'use strict';
 
 const tokens = require('../login/tokens');
-const config = require('../config').config;
+const {config} = require('../config');
 
 /**
  * Express middleware that checks if the user is authenticated.
