@@ -9,8 +9,7 @@ const semver = require('semver');
 
 // Ours
 const bundles = require('../bundle-manager');
-const config = require('../config').config;
-const filteredConfig = require('../config').filteredConfig;
+const {config, filteredConfig} = require('../config');
 const ravenConfig = require('../util/raven-config');
 
 /**
@@ -65,7 +64,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 
 			if (opts.standalone) {
 				// Load the socket
-				scripts.push('<script src="/bower_components/cookies-js/dist/cookies.min.js"></script>');
+				scripts.push('<script src="/node_modules/cookies-js/dist/cookies.min.js"></script>');
 				scripts.push('<script src="/socket.io/socket.io.js"></script>');
 				scripts.push(`<script>
 					const params = new URLSearchParams(location.search);
@@ -84,7 +83,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 			// Likewise, they both need the contentWindow portion of the iframeResizer.
 			// We put this at the start and make it async so it loads ASAP.
 			if (!opts.fullbleed) {
-				scripts.unshift('<script async src="/bower_components/iframe-resizer/js/iframeResizer.contentWindow.js">' +
+				scripts.unshift('<script async src="/node_modules/iframe-resizer/js/iframeResizer.contentWindow.min.js">' +
 					'</script>');
 			}
 
@@ -101,6 +100,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 				} else {
 					styles.push('<link rel="stylesheet" href="/dashboard/css/panel-defaults.css">');
 				}
+
 				scripts.push('<script async src="/dashboard/js/dialog_opener.js"></script>');
 			} else if (resourceType === 'dialog') {
 				styles.push('<link rel="stylesheet" href="/dashboard/css/dialog-defaults.css">');
@@ -108,7 +108,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 		} else if (resourceType === 'graphic') {
 			if (config.sentry && config.sentry.enabled) {
 				scripts.unshift(
-					'<script src="/bower_components/raven-js/dist/raven.min.js"></script>',
+					'<script src="/node_modules/raven-js/dist/raven.min.js"></script>',
 					`<script>
 						Raven.config('${config.sentry.publicDsn}', ${JSON.stringify(ravenConfig)}).install();
 						window.addEventListener('unhandledrejection', function (err) {
@@ -124,7 +124,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 			// If login security is enabled, graphics must provide a valid token when they connect to the socket
 			// They also need cookies-js and a helper to manage URL query strings
 			if (config.login.enabled) {
-				scripts.push('<script src="/bower_components/cookies-js/dist/cookies.min.js"></script>');
+				scripts.push('<script src="/node_modules/cookies-js/dist/cookies.min.js"></script>');
 				scripts.push('<script src="/login/QueryString.js"></script>');
 				scripts.push(`<script>
 					window.token = qs['key'] || Cookies.get('socketToken');
@@ -136,7 +136,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 
 			// If this bundle has sounds, inject SoundJS
 			if (opts.sound) {
-				scripts.push('<script src="/bower_components/SoundJS/lib/soundjs.min.js"></script>');
+				scripts.push('<script src="/node_modules/soundjs/lib/soundjs.min.js"></script>');
 			}
 
 			// Graphics must include the API script themselves before attempting to make an instance of it
@@ -159,7 +159,8 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 		}
 
 		// Inject the scripts required for singleInstance behavior, if requested.
-		if (resourceType === 'graphic') {
+		if (resourceType === 'graphic' &&
+			!(pathOrHtml.endsWith('busy.html') || pathOrHtml.endsWith('killed.html'))) {
 			scripts.push('<script src="/instance/client_registration.js"></script>');
 		}
 
@@ -174,15 +175,14 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 			$('body').append(scripts);
 		}
 
-		// Inject the needed stylesheets, if any.
-		// If there are <script> tags in the head, inject before them.
-		// Otherwise, add at the end of the <head>.
+		// Prepend our styles before the first one.
+		// If there are no styles, put our styles at the end of <head>.
 		if (styles.length > 0) {
 			styles = styles.join('\n');
 
-			const headScripts = $('head').find('script');
-			if (headScripts.length > 0) {
-				headScripts.first().before(styles);
+			const headStyles = $('head').find('style, link[rel="stylesheet"]');
+			if (headStyles.length > 0) {
+				headStyles.first().before(styles);
 			} else {
 				$('head').append(styles);
 			}
