@@ -5,7 +5,7 @@ const bundleManager = require('../bundle-manager');
 const configHelper = require('../config');
 const pjson = require('../../package.json');
 
-const config = configHelper.config;
+const {config} = configHelper;
 const ravenConfigExtra = {};
 const ravenConfig = {
 	release: pjson.version,
