@@ -5,5 +5,5 @@ const replace = require('replace-in-file');
 replace.sync({
 	files: 'instrumented/**/*.js',
 	from: /new Function\('return this'\)/g,
-	to: `new Function('return window.top')`
+	to: 'new Function(\'return window.top\')'
 });
