@@ -1,60 +1,48 @@
-'use strict';
-
 const path = require('path');
-const config = require(path.resolve(__dirname, '../../lib/config')).config;
-const bundleName = 'test-bundle';
 
-function rootUrl() {
-	return `http://localhost:${process.env.NODECG_TEST_PORT}/`;
-}
+export const bundleName = () => 'test-bundle';
 
-function loginUrl() {
-	return `${rootUrl()}login/`;
-}
+export const replicantsRoot = () => path.join(process.env.NODECG_ROOT, 'db/replicants');
 
-function dashboardUrl() {
-	return `${rootUrl()}dashboard/`;
-}
+export const assetsRoot = () => path.join(process.env.NODECG_ROOT, 'assets');
 
-function testBundleRoot() {
-	return `${rootUrl()}bundles/${bundleName}/`;
-}
-
-module.exports = {
-	BUNDLE_NAME: bundleName,
-	CONFIG: config,
-	REPLICANTS_ROOT: path.join(process.env.NODECG_ROOT, 'db/replicants'),
-	ASSETS_ROOT: path.join(process.env.NODECG_ROOT, 'assets'),
-
-	get ROOT_URL() {
-		return rootUrl();
-	},
-
-	get LOGIN_URL() {
-		return loginUrl();
-	},
-
-	get DASHBOARD_URL() {
-		return dashboardUrl();
-	},
-
-	get TEST_PANEL_URL() {
-		return `${testBundleRoot()}dashboard/panel.html`;
-	},
-
-	get BUNDLE_BOWER_COMPONENTS_URL() {
-		return `${testBundleRoot()}bower_components/`;
-	},
-
-	get BUNDLE_NODE_MODULES_URL() {
-		return `${testBundleRoot()}node_modules/`;
-	},
-
-	get GRAPHIC_URL() {
-		return `${testBundleRoot()}graphics/`;
-	},
-
-	get SINGLE_INSTANCE_URL() {
-		return `${testBundleRoot()}graphics/single_instance.html`;
+export const rootUrl = () => {
+	const {NODECG_TEST_PORT} = process.env;
+	if (!NODECG_TEST_PORT) {
+		throw new Error('NODECG_TEST_PORT is missing. Is NodeCG initialized yet?');
 	}
+
+	return `http://localhost:${NODECG_TEST_PORT}/`;
+};
+
+export const loginUrl = () => {
+	return `${rootUrl()}login/`;
+};
+
+export const dashboardUrl = () => {
+	return `${rootUrl()}dashboard/`;
+};
+
+export const testBundleRoot = () => {
+	return `${rootUrl()}bundles/${bundleName()}/`;
+};
+
+export const testPanelUrl = () => {
+	return `${testBundleRoot()}dashboard/panel.html`;
+};
+
+export const bundleBowerComponentsUrl = () => {
+	return `${testBundleRoot()}bower_components/`;
+};
+
+export const bundleNodeModulesUrl = () => {
+	return `${testBundleRoot()}node_modules/`;
+};
+
+export const graphicUrl = () => {
+	return `${testBundleRoot()}graphics/`;
+};
+
+export const singleInstanceUrl = () => {
+	return `${testBundleRoot()}graphics/single_instance.html`;
 };
